% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorMethods.R
\name{dimensions}
\alias{dimensions}
\alias{accessors}
\alias{dimensions,MFCLBase-method}
\alias{dimensions<-}
\alias{dimensions<-,MFCLBase,numeric-method}
\alias{range,MFCLBase-method}
\alias{range<-,MFCLBase,missing,numeric-method}
\alias{n_regions}
\alias{n_regions,MFCLFrqStats-method}
\alias{n_regions<-}
\alias{n_regions<-,MFCLFrqStats,numeric-method}
\alias{n_fisheries}
\alias{n_fisheries,MFCLFrqStats-method}
\alias{n_fisheries<-}
\alias{n_fisheries<-,MFCLFrqStats,numeric-method}
\alias{n_tag_groups}
\alias{n_tag_groups,MFCLFrqStats-method}
\alias{n_tag_groups<-}
\alias{n_tag_groups<-,MFCLFrqStats,numeric-method}
\alias{n_recs_yr}
\alias{n_recs_yr,MFCLFrqStats-method}
\alias{n_recs_yr<-}
\alias{n_recs_yr<-,MFCLFrqStats,numeric-method}
\alias{rec_month}
\alias{rec_month,MFCLFrqStats-method}
\alias{rec_month<-}
\alias{rec_month<-,MFCLFrqStats,numeric-method}
\alias{generic_diffusion}
\alias{generic_diffusion,MFCLFrqStats-method}
\alias{generic_diffusion<-}
\alias{generic_diffusion<-,MFCLFrqStats,logical-method}
\alias{frq_age_len}
\alias{frq_age_len,MFCLFrqStats-method}
\alias{frq_age_len<-}
\alias{frq_age_len<-,MFCLFrqStats,logical-method}
\alias{frq_version}
\alias{frq_version,MFCLFrqStats-method}
\alias{frq_version<-}
\alias{frq_version<-,MFCLFrqStats,numeric-method}
\alias{region_size}
\alias{region_size,MFCLFrqStats-method}
\alias{region_size<-}
\alias{region_size<-,MFCLFrqStats,FLQuant-method}
\alias{region_fish}
\alias{region_fish,MFCLFrqStats-method}
\alias{region_fish<-}
\alias{region_fish<-,MFCLFrqStats,FLQuant-method}
\alias{move_matrix}
\alias{move_matrix,MFCLFrqStats-method}
\alias{move_matrix<-}
\alias{move_matrix<-,MFCLFrqStats,matrix-method}
\alias{data_flags}
\alias{data_flags,MFCLFrqStats-method}
\alias{data_flags<-}
\alias{data_flags<-,MFCLFrqStats,matrix-method}
\alias{season_flags}
\alias{season_flags,MFCLFrqStats-method}
\alias{season_flags<-}
\alias{season_flags<-,MFCLFrqStats,matrix-method}
\alias{n_move_yr}
\alias{n_move_yr,MFCLFrqStats-method}
\alias{n_move_yr<-}
\alias{n_move_yr<-,MFCLFrqStats,numeric-method}
\alias{move_weeks}
\alias{move_weeks,MFCLFrqStats-method}
\alias{move_weeks<-}
\alias{move_weeks<-,MFCLFrqStats,numeric-method}
\alias{range,MFCLFrqStats-method}
\alias{range<-,MFCLFrqStats,missing,numeric-method}
\alias{lf_range}
\alias{lf_range,MFCLLenFreq-method}
\alias{lf_range<-}
\alias{lf_range<-,MFCLLenFreq,numeric-method}
\alias{age_nage}
\alias{age_nage,MFCLLenFreq-method}
\alias{age_nage<-}
\alias{age_nage<-,MFCLLenFreq,numeric-method}
\alias{freq}
\alias{freq,MFCLLenFreq-method}
\alias{freq<-}
\alias{freq<-,MFCLLenFreq,data.frame-method}
\alias{cateffpen}
\alias{cateffpen,MFCLLenFreq2-method}
\alias{cateffpen<-}
\alias{cateffpen<-,MFCLLenFreq2,data.frame-method}
\alias{lnfrq}
\alias{lnfrq,MFCLLenFreq2-method}
\alias{lnfrq<-}
\alias{lnfrq<-,MFCLLenFreq2,data.frame-method}
\alias{wtfrq}
\alias{wtfrq,MFCLLenFreq2-method}
\alias{wtfrq<-}
\alias{wtfrq<-,MFCLLenFreq2,data.frame-method}
\alias{freq,MFCLLenFreq2-method}
\alias{freq<-,MFCLLenFreq2,data.frame-method}
\alias{m,MFCLBiol-method}
\alias{m<-,MFCLBiol,numeric-method}
\alias{m_devs_age}
\alias{m_devs_age,MFCLBiol-method}
\alias{m_devs_age<-}
\alias{m_devs_age<-,MFCLBiol,FLQuant-method}
\alias{log_m}
\alias{log_m,MFCLBiol-method}
\alias{log_m<-}
\alias{log_m<-,MFCLBiol,FLQuant-method}
\alias{mat,MFCLBiol-method}
\alias{mat<-,MFCLBiol,FLQuant-method}
\alias{mat_at_length}
\alias{mat_at_length,MFCLBiol-method}
\alias{mat_at_length<-}
\alias{mat_at_length<-,MFCLBiol,numeric-method}
\alias{growth}
\alias{growth,MFCLBiol-method}
\alias{growth<-}
\alias{growth<-,MFCLBiol,array-method}
\alias{richards}
\alias{richards,MFCLBiol-method}
\alias{richards<-}
\alias{richards<-,MFCLBiol,numeric-method}
\alias{growth_var_pars}
\alias{growth_var_pars,MFCLBiol-method}
\alias{growth_var_pars<-}
\alias{growth_var_pars<-,MFCLBiol,array-method}
\alias{n_mean_constraints}
\alias{n_mean_constraints,MFCLBiol-method}
\alias{n_mean_constraints<-}
\alias{n_mean_constraints<-,MFCLBiol,numeric-method}
\alias{growth_devs_age}
\alias{growth_devs_age,MFCLBiol-method}
\alias{growth_devs_age<-}
\alias{growth_devs_age<-,MFCLBiol,FLQuant-method}
\alias{growth_curve_devs}
\alias{growth_curve_devs,MFCLBiol-method}
\alias{growth_curve_devs<-}
\alias{growth_curve_devs<-,MFCLBiol,FLQuant-method}
\alias{growth_devs_cohort}
\alias{growth_devs_cohort,MFCLBiol-method}
\alias{growth_devs_cohort<-}
\alias{growth_devs_cohort<-,MFCLBiol,FLCohort-method}
\alias{season_growth_pars}
\alias{season_growth_pars,MFCLBiol-method}
\alias{season_growth_pars<-}
\alias{season_growth_pars<-,MFCLBiol,numeric-method}
\alias{len_bias_pars}
\alias{len_bias_pars,MFCLBiol-method}
\alias{len_bias_pars<-}
\alias{len_bias_pars<-,MFCLBiol,numeric-method}
\alias{common_len_bias_pars}
\alias{common_len_bias_pars,MFCLBiol-method}
\alias{common_len_bias_pars<-}
\alias{common_len_bias_pars<-,MFCLBiol,numeric-method}
\alias{common_len_bias_coffs}
\alias{common_len_bias_coffs,MFCLBiol-method}
\alias{common_len_bias_coffs<-}
\alias{common_len_bias_coffs<-,MFCLBiol,numeric-method}
\alias{dimensions,MFCLBiol-method}
\alias{dimensions<-,MFCLBiol,numeric-method}
\alias{range,MFCLBiol-method}
\alias{range<-,MFCLBiol,missing,numeric-method}
\alias{flags}
\alias{flags,MFCLFlags-method}
\alias{flags<-}
\alias{flags<-,MFCLFlags,data.frame-method}
\alias{unused}
\alias{unused,MFCLFlags-method}
\alias{unused<-}
\alias{unused<-,MFCLFlags,list-method}
\alias{tag_shed_rate}
\alias{tag_shed_rate,MFCLTagRep-method}
\alias{tag_shed_rate<-}
\alias{tag_shed_rate<-,MFCLTagRep,numeric-method}
\alias{tag_fish_rep_rate}
\alias{tag_fish_rep_rate,MFCLTagRep-method}
\alias{tag_fish_rep_rate<-}
\alias{tag_fish_rep_rate<-,MFCLTagRep,array-method}
\alias{tag_fish_rep_grp}
\alias{tag_fish_rep_grp,MFCLTagRep-method}
\alias{tag_fish_rep_grp<-}
\alias{tag_fish_rep_grp<-,MFCLTagRep,array-method}
\alias{tag_fish_rep_flags}
\alias{tag_fish_rep_flags,MFCLTagRep-method}
\alias{tag_fish_rep_flags<-}
\alias{tag_fish_rep_flags<-,MFCLTagRep,array-method}
\alias{tag_fish_rep_target}
\alias{tag_fish_rep_target,MFCLTagRep-method}
\alias{tag_fish_rep_target<-}
\alias{tag_fish_rep_target<-,MFCLTagRep,array-method}
\alias{tag_fish_rep_pen}
\alias{tag_fish_rep_pen,MFCLTagRep-method}
\alias{tag_fish_rep_pen<-}
\alias{tag_fish_rep_pen<-,MFCLTagRep,array-method}
\alias{rep_rate_dev_coffs}
\alias{rep_rate_dev_coffs,MFCLTagRep-method}
\alias{rep_rate_dev_coffs<-}
\alias{rep_rate_dev_coffs<-,MFCLTagRep,list-method}
\alias{range,MFCLTagRep-method}
\alias{range<-,MFCLTagRep,missing,numeric-method}
\alias{rec_init_pop_diff}
\alias{rec_init_pop_diff,MFCLRec-method}
\alias{rec_init_pop_diff<-}
\alias{rec_init_pop_diff<-,MFCLRec,numeric-method}
\alias{rec_times}
\alias{rec_times,MFCLRec-method}
\alias{rec_times<-}
\alias{rec_times<-,MFCLRec,numeric-method}
\alias{rel_rec}
\alias{rel_rec,MFCLRec-method}
\alias{rel_rec<-}
\alias{rel_rec<-,MFCLRec,FLQuant-method}
\alias{tot_pop}
\alias{tot_pop,MFCLRec-method}
\alias{tot_pop<-}
\alias{tot_pop<-,MFCLRec,numeric-method}
\alias{tot_pop_implicit}
\alias{tot_pop_implicit,MFCLRec-method}
\alias{tot_pop_implicit<-}
\alias{tot_pop_implicit<-,MFCLRec,numeric-method}
\alias{rel_ini_pop}
\alias{rel_ini_pop,MFCLRec-method}
\alias{rel_ini_pop<-}
\alias{rel_ini_pop<-,MFCLRec,array-method}
\alias{orth_coffs}
\alias{orth_coffs,MFCLRec-method}
\alias{orth_coffs<-}
\alias{orth_coffs<-,MFCLRec,matrix-method}
\alias{new_orth_coffs}
\alias{new_orth_coffs,MFCLRec-method}
\alias{new_orth_coffs<-}
\alias{new_orth_coffs<-,MFCLRec,numeric-method}
\alias{rec_orthogonal}
\alias{rec_orthogonal,MFCLRec-method}
\alias{rec_orthogonal<-}
\alias{rec_orthogonal<-,MFCLRec,FLQuant-method}
\alias{rec_standard_dim}
\alias{rec_standard_dim,MFCLRec-method}
\alias{rec_standard_dim<-}
\alias{rec_standard_dim<-,MFCLRec,numeric-method}
\alias{rec_standard}
\alias{rec_standard,MFCLRec-method}
\alias{rec_standard<-}
\alias{rec_standard<-,MFCLRec,FLQuant-method}
\alias{annual_rel_rec_coffs}
\alias{annual_rel_rec_coffs,MFCLRec-method}
\alias{annual_rel_rec_coffs<-}
\alias{annual_rel_rec_coffs<-,MFCLRec,matrix-method}
\alias{range,MFCLRec-method}
\alias{range<-,MFCLRec,missing,numeric-method}
\alias{control_flags}
\alias{control_flags,MFCLRegion-method}
\alias{control_flags<-}
\alias{control_flags<-,MFCLRegion,matrix-method}
\alias{move_map}
\alias{move_map,MFCLRegion-method}
\alias{move_map<-}
\alias{move_map<-,MFCLRegion,numeric-method}
\alias{diff_coffs}
\alias{diff_coffs,MFCLRegion-method}
\alias{diff_coffs<-}
\alias{diff_coffs<-,MFCLRegion,matrix-method}
\alias{xdiff_coffs}
\alias{xdiff_coffs,MFCLRegion-method}
\alias{xdiff_coffs<-}
\alias{xdiff_coffs<-,MFCLRegion,matrix-method}
\alias{y1diff_coffs}
\alias{y1diff_coffs,MFCLRegion-method}
\alias{y1diff_coffs<-}
\alias{y1diff_coffs<-,MFCLRegion,matrix-method}
\alias{y2diff_coffs}
\alias{y2diff_coffs,MFCLRegion-method}
\alias{y2diff_coffs<-}
\alias{y2diff_coffs<-,MFCLRegion,matrix-method}
\alias{zdiff_coffs}
\alias{zdiff_coffs,MFCLRegion-method}
\alias{zdiff_coffs<-}
\alias{zdiff_coffs<-,MFCLRegion,matrix-method}
\alias{diff_coffs_mat}
\alias{diff_coffs_mat,MFCLRegion-method}
\alias{diff_coffs_mat<-}
\alias{diff_coffs_mat<-,MFCLRegion,matrix-method}
\alias{diff_coffs_age_ssn}
\alias{diff_coffs_age_ssn,MFCLRegion-method}
\alias{diff_coffs_age_ssn<-}
\alias{diff_coffs_age_ssn<-,MFCLRegion,array-method}
\alias{diff_coffs_age_period}
\alias{diff_coffs_age_period,MFCLRegion-method}
\alias{diff_coffs_age_period<-}
\alias{diff_coffs_age_period<-,MFCLRegion,array-method}
\alias{diff_coffs_age}
\alias{diff_coffs_age,MFCLRegion-method}
\alias{diff_coffs_age<-}
\alias{diff_coffs_age<-,MFCLRegion,array-method}
\alias{diff_coffs_nl}
\alias{diff_coffs_nl,MFCLRegion-method}
\alias{diff_coffs_nl<-}
\alias{diff_coffs_nl<-,MFCLRegion,array-method}
\alias{diff_coffs_priors}
\alias{diff_coffs_priors,MFCLRegion-method}
\alias{diff_coffs_priors<-}
\alias{diff_coffs_priors<-,MFCLRegion,array-method}
\alias{diff_coffs_age_priors}
\alias{diff_coffs_age_priors,MFCLRegion-method}
\alias{diff_coffs_age_priors<-}
\alias{diff_coffs_age_priors<-,MFCLRegion,array-method}
\alias{diff_coffs_nl_priors}
\alias{diff_coffs_nl_priors,MFCLRegion-method}
\alias{diff_coffs_nl_priors<-}
\alias{diff_coffs_nl_priors<-,MFCLRegion,array-method}
\alias{region_rec_var}
\alias{region_rec_var,MFCLRegion-method}
\alias{region_rec_var<-}
\alias{region_rec_var<-,MFCLRegion,FLQuant-method}
\alias{region_pars}
\alias{region_pars,MFCLRegion-method}
\alias{region_pars<-}
\alias{region_pars<-,MFCLRegion,matrix-method}
\alias{range,MFCLRegion-method}
\alias{range<-,MFCLRegion,missing,numeric-method}
\alias{availability_coffs}
\alias{availability_coffs,MFCLSel-method}
\alias{availability_coffs<-}
\alias{availability_coffs<-,MFCLSel,FLQuant-method}
\alias{fishery_sel}
\alias{fishery_sel,MFCLSel-method}
\alias{fishery_sel<-}
\alias{fishery_sel<-,MFCLSel,FLQuant-method}
\alias{fishery_sel_age_comp}
\alias{fishery_sel_age_comp,MFCLSel-method}
\alias{fishery_sel_age_comp<-}
\alias{fishery_sel_age_comp<-,MFCLSel,FLQuant-method}
\alias{av_q_coffs}
\alias{av_q_coffs,MFCLSel-method}
\alias{av_q_coffs<-}
\alias{av_q_coffs<-,MFCLSel,FLQuant-method}
\alias{ini_q_coffs}
\alias{ini_q_coffs,MFCLSel-method}
\alias{ini_q_coffs<-}
\alias{ini_q_coffs<-,MFCLSel,FLQuant-method}
\alias{q0_miss}
\alias{q0_miss,MFCLSel-method}
\alias{q0_miss<-}
\alias{q0_miss<-,MFCLSel,FLQuant-method}
\alias{q_dev_coffs}
\alias{q_dev_coffs,MFCLSel-method}
\alias{q_dev_coffs<-}
\alias{q_dev_coffs<-,MFCLSel,list-method}
\alias{effort_dev_coffs}
\alias{effort_dev_coffs,MFCLSel,missing-method}
\alias{effort_dev_coffs<-}
\alias{effort_dev_coffs<-,MFCLSel,list-method}
\alias{catch_dev_coffs}
\alias{catch_dev_coffs,MFCLSel-method}
\alias{catch_dev_coffs<-}
\alias{catch_dev_coffs<-,MFCLSel,list-method}
\alias{catch_dev_coffs_flag}
\alias{catch_dev_coffs_flag,MFCLSel-method}
\alias{catch_dev_coffs_flag<-}
\alias{catch_dev_coffs_flag<-,MFCLSel,numeric-method}
\alias{sel_dev_corr}
\alias{sel_dev_corr,MFCLSel-method}
\alias{sel_dev_corr<-}
\alias{sel_dev_corr<-,MFCLSel,FLQuant-method}
\alias{sel_dev_coffs}
\alias{sel_dev_coffs,MFCLSel-method}
\alias{sel_dev_coffs<-}
\alias{sel_dev_coffs<-,MFCLSel,matrix-method}
\alias{sel_dev_coffs2}
\alias{sel_dev_coffs2,MFCLSel-method}
\alias{sel_dev_coffs2<-}
\alias{sel_dev_coffs2<-,MFCLSel,list-method}
\alias{season_q_pars}
\alias{season_q_pars,MFCLSel-method}
\alias{season_q_pars<-}
\alias{season_q_pars<-,MFCLSel,matrix-method}
\alias{fish_params}
\alias{fish_params,MFCLSel-method}
\alias{fish_params<-}
\alias{fish_params<-,MFCLSel,matrix-method}
\alias{spp_params}
\alias{spp_params,MFCLSel-method}
\alias{spp_params<-}
\alias{spp_params<-,MFCLSel,matrix-method}
\alias{range,MFCLSel-method}
\alias{range<-,MFCLSel,missing,numeric-method}
\alias{fm_level_devs}
\alias{fm_level_devs,MFCLParBits-method}
\alias{fm_level_devs<-}
\alias{fm_level_devs<-,MFCLParBits,character-method}
\alias{fm_level_regression_pars}
\alias{fm_level_regression_pars,MFCLParBits-method}
\alias{fm_level_regression_pars<-}
\alias{fm_level_regression_pars<-,MFCLParBits,matrix-method}
\alias{obj_fun}
\alias{obj_fun,MFCLParBits-method}
\alias{obj_fun<-}
\alias{obj_fun<-,MFCLParBits,numeric-method}
\alias{n_pars}
\alias{n_pars,MFCLParBits-method}
\alias{n_pars<-}
\alias{n_pars<-,MFCLParBits,numeric-method}
\alias{tag_lik}
\alias{tag_lik,MFCLParBits-method}
\alias{tag_lik<-}
\alias{tag_lik<-,MFCLParBits,numeric-method}
\alias{mn_len_pen}
\alias{mn_len_pen,MFCLParBits-method}
\alias{mn_len_pen<-}
\alias{mn_len_pen<-,MFCLParBits,numeric-method}
\alias{max_grad}
\alias{max_grad,MFCLParBits-method}
\alias{max_grad<-}
\alias{max_grad<-,MFCLParBits,numeric-method}
\alias{av_fish_mort_inst}
\alias{av_fish_mort_inst,MFCLParBits-method}
\alias{av_fish_mort_inst<-}
\alias{av_fish_mort_inst<-,MFCLParBits,numeric-method}
\alias{av_fish_mort_year}
\alias{av_fish_mort_year,MFCLParBits-method}
\alias{av_fish_mort_year<-}
\alias{av_fish_mort_year<-,MFCLParBits,numeric-method}
\alias{av_fish_mort_age}
\alias{av_fish_mort_age,MFCLParBits-method}
\alias{av_fish_mort_age<-}
\alias{av_fish_mort_age<-,MFCLParBits,numeric-method}
\alias{logistic_normal_params}
\alias{logistic_normal_params,MFCLParBits-method}
\alias{logistic_normal_params<-}
\alias{logistic_normal_params<-,MFCLParBits,character-method}
\alias{lagrangian}
\alias{lagrangian,MFCLParBits-method}
\alias{lagrangian<-}
\alias{lagrangian<-,MFCLParBits,character-method}
\alias{kludged_eq_coffs}
\alias{kludged_eq_coffs,MFCLParBits-method}
\alias{kludged_eq_coffs<-}
\alias{kludged_eq_coffs<-,MFCLParBits,array-method}
\alias{kludged_eq_level_coffs}
\alias{kludged_eq_level_coffs,MFCLParBits-method}
\alias{kludged_eq_level_coffs<-}
\alias{kludged_eq_level_coffs<-,MFCLParBits,numeric-method}
\alias{first_year}
\alias{first_year,MFCLParBits-method}
\alias{first_year<-}
\alias{first_year<-,MFCLParBits,numeric-method}
\alias{range,MFCLParBits-method}
\alias{range<-,MFCLParBits,missing,numeric-method}
\alias{age_pars}
\alias{age_pars,MFCLIniBits-method}
\alias{age_pars<-}
\alias{age_pars<-,MFCLIniBits,matrix-method}
\alias{rec_dist}
\alias{rec_dist,MFCLIniBits-method}
\alias{rec_dist<-}
\alias{rec_dist<-,MFCLIniBits,numeric-method}
\alias{lw_params}
\alias{lw_params,MFCLIni-method}
\alias{lw_params<-}
\alias{lw_params<-,MFCLIniBits,numeric-method}
\alias{sv,MFCLIniBits,missing-method}
\alias{sv<-}
\alias{sv<-,MFCLIniBits,numeric-method}
\alias{sd_length_at_age}
\alias{sd_length_at_age,MFCLIniBits-method}
\alias{sd_length_at_age<-}
\alias{sd_length_at_age<-,MFCLIniBits,numeric-method}
\alias{sd_length_dep}
\alias{sd_length_dep,MFCLIniBits-method}
\alias{sd_length_dep<-}
\alias{sd_length_dep<-,MFCLIniBits,numeric-method}
\alias{tot_pop,MFCLIniBits-method}
\alias{tot_pop<-,MFCLIniBits,numeric-method}
\alias{tag_flags}
\alias{tag_flags,MFCLIniBits-method}
\alias{tag_flags<-}
\alias{tag_flags<-,MFCLIniBits,matrix-method}
\alias{fishery_realizations}
\alias{fishery_realizations,MFCLRep-method}
\alias{fishery_realizations<-}
\alias{fishery_realizations<-,MFCLRep,FLQuant-method}
\alias{mean_laa}
\alias{mean_laa,MFCLRep-method}
\alias{mean_laa<-}
\alias{mean_laa<-,MFCLRep,FLQuant-method}
\alias{mean_waa}
\alias{mean_waa,MFCLRep-method}
\alias{mean_waa<-}
\alias{mean_waa<-,MFCLRep,FLQuant-method}
\alias{sd_laa}
\alias{sd_laa,MFCLRep-method}
\alias{sd_laa<-}
\alias{sd_laa<-,MFCLRep,FLQuant-method}
\alias{m_at_age}
\alias{m_at_age,MFCLRep-method}
\alias{m_at_age<-}
\alias{m_at_age<-,MFCLRep,FLQuant-method}
\alias{sel}
\alias{sel,MFCLRep-method}
\alias{sel<-}
\alias{sel<-,MFCLRep,FLQuant-method}
\alias{q_fishery}
\alias{q_fishery,MFCLRep-method}
\alias{q_fishery<-}
\alias{q_fishery<-,MFCLRep,FLQuant-method}
\alias{q_effdev}
\alias{q_effdev,MFCLRep-method}
\alias{q_effdev<-}
\alias{q_effdev<-,MFCLRep,FLQuant-method}
\alias{fm}
\alias{fm,MFCLRep-method}
\alias{fm<-}
\alias{fm<-,MFCLRep,FLQuant-method}
\alias{fm_aggregated}
\alias{fm_aggregated,MFCLRep-method}
\alias{fm_aggregated<-}
\alias{fm_aggregated<-,MFCLRep,FLQuant-method}
\alias{popN}
\alias{popN,MFCLRep-method}
\alias{popN<-}
\alias{popN<-,MFCLRep,FLQuant-method}
\alias{rec_region}
\alias{rec_region,MFCLRep-method}
\alias{rec_region<-}
\alias{rec_region<-,MFCLRep,FLQuant-method}
\alias{totalBiomass}
\alias{totalBiomass,MFCLRep-method}
\alias{totalBiomass<-}
\alias{totalBiomass<-,MFCLRep,FLQuant-method}
\alias{totalBiomass_nofish}
\alias{totalBiomass_nofish,MFCLRep-method}
\alias{totalBiomass_nofish<-}
\alias{totalBiomass_nofish<-,MFCLRep,FLQuant-method}
\alias{adultBiomass}
\alias{adultBiomass,MFCLRep-method}
\alias{adultBiomass<-}
\alias{adultBiomass<-,MFCLRep,FLQuant-method}
\alias{adultBiomass_nofish}
\alias{adultBiomass_nofish,MFCLRep-method}
\alias{adultBiomass_nofish<-}
\alias{adultBiomass_nofish<-,MFCLRep,FLQuant-method}
\alias{vulnBiomass}
\alias{vulnBiomass,MFCLRep-method}
\alias{vulnBiomass<-}
\alias{vulnBiomass<-,MFCLRep,FLQuant-method}
\alias{srr}
\alias{srr,MFCLRep-method}
\alias{srr<-}
\alias{srr<-,MFCLRep,FLPar-method}
\alias{eq_ssb}
\alias{eq_ssb,MFCLRep-method}
\alias{eq_ssb<-}
\alias{eq_ssb<-,MFCLRep,FLQuant-method}
\alias{eq_ssb_obs}
\alias{eq_ssb_obs,MFCLRep-method}
\alias{eq_ssb_obs<-}
\alias{eq_ssb_obs<-,MFCLRep,FLQuant-method}
\alias{eq_rec}
\alias{eq_rec,MFCLRep-method}
\alias{eq_rec<-}
\alias{eq_rec<-,MFCLRep,FLQuant-method}
\alias{eq_rec_obs}
\alias{eq_rec_obs,MFCLRep-method}
\alias{eq_rec_obs<-}
\alias{eq_rec_obs<-,MFCLRep,FLQuant-method}
\alias{catch_obs}
\alias{catch_obs,MFCLRep-method}
\alias{catch_obs<-}
\alias{catch_obs<-,MFCLRep,FLQuant-method}
\alias{catch_pred}
\alias{catch_pred,MFCLRep-method}
\alias{catch_pred<-}
\alias{catch_pred<-,MFCLRep,FLQuant-method}
\alias{cpue_obs}
\alias{cpue_obs,MFCLRep-method}
\alias{cpue_obs<-}
\alias{cpue_obs<-,MFCLRep,FLQuant-method}
\alias{cpue_pred}
\alias{cpue_pred,MFCLRep-method}
\alias{cpue_pred<-}
\alias{cpue_pred<-,MFCLRep,FLQuant-method}
\alias{eq_biomass}
\alias{eq_biomass,MFCLRep-method}
\alias{eq_biomass<-}
\alias{eq_biomass<-,MFCLRep,FLQuant-method}
\alias{eq_yield}
\alias{eq_yield,MFCLRep-method}
\alias{eq_yield<-}
\alias{eq_yield<-,MFCLRep,FLQuant-method}
\alias{MSY}
\alias{MSY,MFCLRep-method}
\alias{MSY<-}
\alias{MSY<-,MFCLRep,numeric-method}
\alias{FMSY}
\alias{FMSY,MFCLRep-method}
\alias{FMSY<-}
\alias{FMSY<-,MFCLRep,numeric-method}
\alias{BMSY}
\alias{BMSY,MFCLRep-method}
\alias{BMSY<-}
\alias{BMSY<-,MFCLRep,numeric-method}
\alias{FFMSY_ts}
\alias{FFMSY_ts,MFCLRep-method}
\alias{FFMSY_ts<-}
\alias{FFMSY_ts<-,MFCLRep,FLQuant-method}
\alias{ABBMSY_ts}
\alias{ABBMSY_ts,MFCLRep-method}
\alias{ABBMSY_ts<-}
\alias{ABBMSY_ts<-,MFCLRep,FLQuant-method}
\alias{ABBMSY}
\alias{ABBMSY,MFCLRep-method}
\alias{ABBMSY<-}
\alias{ABBMSY<-,MFCLRep,numeric-method}
\alias{TBBMSY}
\alias{TBBMSY,MFCLRep-method}
\alias{TBBMSY<-}
\alias{TBBMSY<-,MFCLRep,numeric-method}
\alias{Fmult}
\alias{Fmult,MFCLRep-method}
\alias{Fmult<-}
\alias{Fmult<-,MFCLRep,numeric-method}
\alias{AggregateF}
\alias{AggregateF,MFCLRep-method}
\alias{AggregateF<-}
\alias{AggregateF<-,MFCLRep,FLQuant-method}
\alias{dimensions,MFCLRep-method}
\alias{dimensions<-,MFCLRep,numeric-method}
\alias{range,MFCLRep-method}
\alias{range<-,MFCLRep,missing,numeric-method}
\alias{nyears}
\alias{nyears,MFCLprojControl-method}
\alias{nyears<-}
\alias{nyears<-,MFCLprojControl,numeric-method}
\alias{nsims}
\alias{nsims,MFCLprojControl-method}
\alias{nsims<-}
\alias{nsims<-,MFCLprojControl,numeric-method}
\alias{avyrs}
\alias{avyrs,MFCLprojControl-method}
\alias{avyrs<-}
\alias{avyrs<-,MFCLprojControl,character-method}
\alias{fprojyr}
\alias{fprojyr,MFCLprojControl-method}
\alias{fprojyr<-}
\alias{fprojyr<-,MFCLprojControl,numeric-method}
\alias{controls}
\alias{controls,MFCLprojControl-method}
\alias{controls<-}
\alias{controls<-,MFCLprojControl,data.frame-method}
\alias{caeff}
\alias{caeff,MFCLprojControl-method}
\alias{caeff<-}
\alias{caeff<-,MFCLprojControl,numeric-method}
\alias{scaler}
\alias{scaler,MFCLprojControl-method}
\alias{scaler<-}
\alias{scaler<-,MFCLprojControl,numeric-method}
\alias{ess}
\alias{ess,MFCLprojControl-method}
\alias{ess<-}
\alias{ess<-,MFCLprojControl,data.frame-method}
\alias{range,MFCLFrq-method}
\alias{range<-,MFCLFrq,missing,numeric-method}
\alias{range,MFCLPar-method}
\alias{range<-,MFCLPar,missing,numeric-method}
\alias{dimensions,MFCLPar-method}
\alias{dimensions<-,MFCLPar,numeric-method}
\alias{range,MFCLIni-method}
\alias{range<-,MFCLIni,missing,numeric-method}
\alias{ini_version}
\alias{ini_version,MFCLIniBits-method}
\alias{ini_version<-}
\alias{ini_version<-,MFCLIniBits,numeric-method}
\alias{region_flags}
\alias{region_flags,MFCLIniBits-method}
\alias{region_flags<-}
\alias{region_flags<-,MFCLIniBits,matrix-method}
\alias{release_groups_proj}
\alias{release_groups_proj,MFCLTagProj-method}
\alias{release_groups_proj<-}
\alias{release_groups_proj<-,MFCLTagProj,numeric-method}
\alias{releases_proj}
\alias{releases_proj,MFCLTagProj-method}
\alias{releases_proj<-}
\alias{releases_proj<-,MFCLTagProj,data.frame-method}
\alias{rep_rate_proj}
\alias{rep_rate_proj,MFCLTagProj-method}
\alias{rep_rate_proj<-}
\alias{rep_rate_proj<-,MFCLTagProj,array-method}
\alias{range,MFCLTagProj-method}
\alias{range<-,MFCLTagProj,missing,numeric-method}
\alias{release_groups}
\alias{release_groups,MFCLTag-method}
\alias{release_groups<-}
\alias{release_groups<-,MFCLTag,numeric-method}
\alias{release_lengths}
\alias{release_lengths,MFCLTag-method}
\alias{release_lengths<-}
\alias{release_lengths<-,MFCLTag,numeric-method}
\alias{recoveries}
\alias{recoveries,MFCLTag-method}
\alias{recoveries<-}
\alias{recoveries<-,MFCLTag,numeric-method}
\alias{releases}
\alias{releases,MFCLTag-method}
\alias{releases<-}
\alias{releases<-,MFCLTag,data.frame-method}
\alias{recaptures}
\alias{recaptures,MFCLTag-method}
\alias{recaptures<-}
\alias{recaptures<-,MFCLTag,data.frame-method}
\alias{range,MFCLTag-method}
\alias{range<-,MFCLTag,missing,numeric-method}
\alias{total_catch}
\alias{total_catch,MFCLCatch-method}
\alias{total_catch<-}
\alias{total_catch<-,MFCLCatch,FLQuant-method}
\alias{fishery_catch}
\alias{fishery_catch,MFCLCatch-method}
\alias{fishery_catch<-}
\alias{fishery_catch<-,MFCLCatch,FLQuant-method}
\alias{range,MFCLCatch-method}
\alias{range<-,MFCLCatch,missing,numeric-method}
\alias{laa}
\alias{laa,MFCLLenFit-method}
\alias{laa<-}
\alias{laa<-,MFCLLenFit,FLQuant-method}
\alias{lenfits}
\alias{lenfits,MFCLLenFit-method}
\alias{lenfits<-}
\alias{lenfits<-,MFCLLenFit,data.frame-method}
\alias{lenagefits}
\alias{lenagefits,MFCLLenFit-method}
\alias{lenagefits<-}
\alias{lenagefits<-,MFCLLenFit,data.frame-method}
\alias{range,MFCLLenFit-method}
\alias{range<-,MFCLLenFit,missing,numeric-method}
\alias{waa}
\alias{waa,MFCLWgtFit-method}
\alias{waa<-}
\alias{waa<-,MFCLWgtFit,FLQuant-method}
\alias{wgtfits}
\alias{wgtfits,MFCLWgtFit-method}
\alias{wgtfits<-}
\alias{wgtfits<-,MFCLWgtFit,data.frame-method}
\alias{wgtagefits}
\alias{wgtagefits,MFCLWgtFit-method}
\alias{wgtagefits<-}
\alias{wgtagefits<-,MFCLWgtFit,data.frame-method}
\alias{range,MFCLWgtFit-method}
\alias{range<-,MFCLWgtFit,missing,numeric-method}
\alias{bh_steep_contrib}
\alias{bh_steep_contrib,MFCLLikelihood-method}
\alias{bh_steep_contrib<-}
\alias{bh_steep_contrib<-,MFCLLikelihood,numeric-method}
\alias{effort_dev_penalty}
\alias{effort_dev_penalty,MFCLLikelihood-method}
\alias{effort_dev_penalty<-}
\alias{effort_dev_penalty<-,MFCLLikelihood,numeric-method}
\alias{q_dev_pen_fish}
\alias{q_dev_pen_fish,MFCLLikelihood-method}
\alias{q_dev_pen_fish<-}
\alias{q_dev_pen_fish<-,MFCLLikelihood,numeric-method}
\alias{q_dev_pen_fish_grp}
\alias{q_dev_pen_fish_grp,MFCLLikelihood-method}
\alias{q_dev_pen_fish_grp<-}
\alias{q_dev_pen_fish_grp<-,MFCLLikelihood,numeric-method}
\alias{total_length_fish}
\alias{total_length_fish,MFCLLikelihood-method}
\alias{total_length_fish<-}
\alias{total_length_fish<-,MFCLLikelihood,numeric-method}
\alias{length_fish}
\alias{length_fish,MFCLLikelihood-method}
\alias{length_fish<-}
\alias{length_fish<-,MFCLLikelihood,list-method}
\alias{total_weight_fish}
\alias{total_weight_fish,MFCLLikelihood-method}
\alias{total_weight_fish<-}
\alias{total_weight_fish<-,MFCLLikelihood,numeric-method}
\alias{weight_fish}
\alias{weight_fish,MFCLLikelihood-method}
\alias{weight_fish<-}
\alias{weight_fish<-,MFCLLikelihood,list-method}
\alias{total_catch_fish}
\alias{total_catch_fish,MFCLLikelihood-method}
\alias{total_catch_fish<-}
\alias{total_catch_fish<-,MFCLLikelihood,numeric-method}
\alias{catch_fish}
\alias{catch_fish,MFCLLikelihood-method}
\alias{catch_fish<-}
\alias{catch_fish<-,MFCLLikelihood,list-method}
\alias{tag_rel_fish}
\alias{tag_rel_fish,MFCLLikelihood-method}
\alias{tag_rel_fish<-}
\alias{tag_rel_fish<-,MFCLLikelihood,list-method}
\alias{survey_index}
\alias{survey_index,MFCLLikelihood-method}
\alias{survey_index<-}
\alias{survey_index<-,MFCLLikelihood,numeric-method}
\alias{age_length}
\alias{age_length,MFCLLikelihood-method}
\alias{age_length<-}
\alias{age_length<-,MFCLLikelihood,numeric-method}
\alias{dimensions,MFCLLikelihood-method}
\alias{dimensions<-,MFCLLikelihood,numeric-method}
\alias{catcheff}
\alias{catcheff,MFCLPseudo-method}
\alias{catcheff<-}
\alias{catcheff<-,MFCLPseudo,data.frame-method}
\alias{l_frq}
\alias{l_frq,MFCLPseudo-method}
\alias{l_frq<-}
\alias{l_frq<-,MFCLPseudo,data.frame-method}
\alias{w_frq}
\alias{w_frq,MFCLPseudo-method}
\alias{w_frq<-}
\alias{w_frq<-,MFCLPseudo,data.frame-method}
\alias{catch_sd}
\alias{catch_sd,MFCLPseudoControl-method}
\alias{catch_sd<-}
\alias{catch_sd<-,MFCLPseudoControl,numeric-method}
\alias{effort_sd}
\alias{effort_sd,MFCLPseudoControl-method}
\alias{effort_sd<-}
\alias{effort_sd<-,MFCLPseudoControl,numeric-method}
\alias{tag_controls}
\alias{tag_controls,MFCLPseudoControl-method}
\alias{tag_controls<-}
\alias{tag_controls<-,MFCLPseudoControl,data.frame-method}
\alias{random_seeds}
\alias{random_seeds,MFCLPseudoControl-method}
\alias{random_seeds<-}
\alias{random_seeds<-,MFCLPseudoControl,numeric-method}
\alias{tag_fish_rep_rate,MFCLPseudoControl-method}
\alias{tag_fish_rep_rate<-,MFCLPseudoControl,numeric-method}
\alias{doitall}
\alias{doitall,MFCLEMControl-method}
\alias{doitall<-}
\alias{doitall<-,MFCLEMControl,function-method}
\alias{tag_fish_rep_rate,MFCLEMControl-method}
\alias{tag_fish_rep_rate<-,MFCLEMControl,numeric-method}
\alias{hcr}
\alias{hcr,MFCLMSEControl-method}
\alias{hcr<-}
\alias{hcr<-,MFCLMSEControl,character-method}
\alias{hcr_params}
\alias{hcr_params,MFCLMSEControl-method}
\alias{hcr_params<-}
\alias{hcr_params<-,MFCLMSEControl,numeric-method}
\alias{itn}
\alias{itn,MFCLMSEControl-method}
\alias{itn<-}
\alias{effort_creep}
\alias{effort_creep,MFCLMSEControl-method}
\alias{effort_creep<-}
\alias{effort_creep<-,MFCLMSEControl,numeric-method}
\alias{effort_creep_fish}
\alias{effort_creep_fish,MFCLMSEControl-method}
\alias{effort_creep_fish<-}
\alias{effort_creep_fish<-,MFCLMSEControl,numeric-method}
\alias{ALK}
\alias{ALK,MFCLALK-method}
\alias{ALK<-}
\alias{ALK<-,MFCLALK,data.frame-method}
\alias{ESS}
\alias{ESS,MFCLALK-method}
\alias{ESS<-}
\alias{ESS<-,MFCLALK,numeric-method}
\alias{range,MFCLALK-method}
\alias{range<-,MFCLALK,missing,numeric-method}
\alias{Eq_calcs}
\alias{Eq_calcs,MFCLEquilibrium-method}
\alias{Eq_calcs<-}
\alias{Eq_calcs<-,MFCLEquilibrium,data.frame-method}
\alias{index,MFCLRegScale-method}
\alias{index<-,MFCLRegScale,matrix-method}
\alias{range,MFCLRegScale-method}
\alias{range<-,MFCLRegScale,missing,numeric-method}
\alias{dimensions,MFCLRegScale-method}
\alias{dimensions<-,MFCLRegScale,numeric-method}
\title{Accessor methods}
\usage{
dimensions(object, ...)

\S4method{dimensions}{MFCLBase}(object)

dimensions(object, ...) <- value

\S4method{dimensions}{MFCLBase,numeric}(object) <- value

\S4method{range}{MFCLBase}(x)

\S4method{range}{MFCLBase,missing,numeric}(x) <- value

n_regions(object, ...)

\S4method{n_regions}{MFCLFrqStats}(object)

n_regions(object, ...) <- value

\S4method{n_regions}{MFCLFrqStats,numeric}(object) <- value

n_fisheries(object, ...)

\S4method{n_fisheries}{MFCLFrqStats}(object)

n_fisheries(object, ...) <- value

\S4method{n_fisheries}{MFCLFrqStats,numeric}(object) <- value

n_tag_groups(object, ...)

\S4method{n_tag_groups}{MFCLFrqStats}(object)

n_tag_groups(object, ...) <- value

\S4method{n_tag_groups}{MFCLFrqStats,numeric}(object) <- value

n_recs_yr(object, ...)

\S4method{n_recs_yr}{MFCLFrqStats}(object)

n_recs_yr(object, ...) <- value

\S4method{n_recs_yr}{MFCLFrqStats,numeric}(object) <- value

rec_month(object, ...)

\S4method{rec_month}{MFCLFrqStats}(object)

rec_month(object, ...) <- value

\S4method{rec_month}{MFCLFrqStats,numeric}(object) <- value

generic_diffusion(object, ...)

\S4method{generic_diffusion}{MFCLFrqStats}(object)

generic_diffusion(object, ...) <- value

\S4method{generic_diffusion}{MFCLFrqStats,logical}(object) <- value

frq_age_len(object, ...)

\S4method{frq_age_len}{MFCLFrqStats}(object)

frq_age_len(object, ...) <- value

\S4method{frq_age_len}{MFCLFrqStats,logical}(object) <- value

frq_version(object, ...)

\S4method{frq_version}{MFCLFrqStats}(object)

frq_version(object, ...) <- value

\S4method{frq_version}{MFCLFrqStats,numeric}(object) <- value

region_size(object, ...)

\S4method{region_size}{MFCLFrqStats}(object)

region_size(object, ...) <- value

\S4method{region_size}{MFCLFrqStats,FLQuant}(object) <- value

region_fish(object, ...)

\S4method{region_fish}{MFCLFrqStats}(object)

region_fish(object, ...) <- value

\S4method{region_fish}{MFCLFrqStats,FLQuant}(object) <- value

move_matrix(object, ...)

\S4method{move_matrix}{MFCLFrqStats}(object)

move_matrix(object, ...) <- value

\S4method{move_matrix}{MFCLFrqStats,matrix}(object) <- value

data_flags(object, ...)

\S4method{data_flags}{MFCLFrqStats}(object)

data_flags(object, ...) <- value

\S4method{data_flags}{MFCLFrqStats,matrix}(object) <- value

season_flags(object, ...)

\S4method{season_flags}{MFCLFrqStats}(object)

season_flags(object, ...) <- value

\S4method{season_flags}{MFCLFrqStats,matrix}(object) <- value

n_move_yr(object, ...)

\S4method{n_move_yr}{MFCLFrqStats}(object)

n_move_yr(object, ...) <- value

\S4method{n_move_yr}{MFCLFrqStats,numeric}(object) <- value

move_weeks(object, ...)

\S4method{move_weeks}{MFCLFrqStats}(object)

move_weeks(object, ...) <- value

\S4method{move_weeks}{MFCLFrqStats,numeric}(object) <- value

\S4method{range}{MFCLFrqStats}(x)

\S4method{range}{MFCLFrqStats,missing,numeric}(x) <- value

lf_range(object, ...)

\S4method{lf_range}{MFCLLenFreq}(object)

lf_range(object, ...) <- value

\S4method{lf_range}{MFCLLenFreq,numeric}(object) <- value

age_nage(object, ...)

\S4method{age_nage}{MFCLLenFreq}(object)

age_nage(object, ...) <- value

\S4method{age_nage}{MFCLLenFreq,numeric}(object) <- value

freq(object, ...)

\S4method{freq}{MFCLLenFreq}(object)

freq(object, ...) <- value

\S4method{freq}{MFCLLenFreq,data.frame}(object) <- value

cateffpen(object, ...)

\S4method{cateffpen}{MFCLLenFreq2}(object)

cateffpen(object, ...) <- value

\S4method{cateffpen}{MFCLLenFreq2,data.frame}(object) <- value

lnfrq(object, ...)

\S4method{lnfrq}{MFCLLenFreq2}(object)

lnfrq(object, ...) <- value

\S4method{lnfrq}{MFCLLenFreq2,data.frame}(object) <- value

wtfrq(object, ...)

\S4method{wtfrq}{MFCLLenFreq2}(object)

wtfrq(object, ...) <- value

\S4method{wtfrq}{MFCLLenFreq2,data.frame}(object) <- value

\S4method{freq}{MFCLLenFreq2}(object)

\S4method{freq}{MFCLLenFreq2,data.frame}(object) <- value

\S4method{m}{MFCLBiol}(object)

\S4method{m}{MFCLBiol,numeric}(object) <- value

m_devs_age(object, ...)

\S4method{m_devs_age}{MFCLBiol}(object)

m_devs_age(object, ...) <- value

\S4method{m_devs_age}{MFCLBiol,FLQuant}(object) <- value

log_m(object, ...)

\S4method{log_m}{MFCLBiol}(object)

log_m(object, ...) <- value

\S4method{log_m}{MFCLBiol,FLQuant}(object) <- value

\S4method{mat}{MFCLBiol}(object)

\S4method{mat}{MFCLBiol,FLQuant}(object) <- value

mat_at_length(object, ...)

\S4method{mat_at_length}{MFCLBiol}(object)

mat_at_length(object, ...) <- value

\S4method{mat_at_length}{MFCLBiol,numeric}(object) <- value

growth(object, ...)

\S4method{growth}{MFCLBiol}(object)

growth(object, ...) <- value

\S4method{growth}{MFCLBiol,array}(object) <- value

richards(object, ...)

\S4method{richards}{MFCLBiol}(object)

richards(object, ...) <- value

\S4method{richards}{MFCLBiol,numeric}(object) <- value

growth_var_pars(object, ...)

\S4method{growth_var_pars}{MFCLBiol}(object)

growth_var_pars(object, ...) <- value

\S4method{growth_var_pars}{MFCLBiol,array}(object) <- value

n_mean_constraints(object, ...)

\S4method{n_mean_constraints}{MFCLBiol}(object)

n_mean_constraints(object, ...) <- value

\S4method{n_mean_constraints}{MFCLBiol,numeric}(object) <- value

growth_devs_age(object, ...)

\S4method{growth_devs_age}{MFCLBiol}(object)

growth_devs_age(object, ...) <- value

\S4method{growth_devs_age}{MFCLBiol,FLQuant}(object) <- value

growth_curve_devs(object, ...)

\S4method{growth_curve_devs}{MFCLBiol}(object)

growth_curve_devs(object, ...) <- value

\S4method{growth_curve_devs}{MFCLBiol,FLQuant}(object) <- value

growth_devs_cohort(object, ...)

\S4method{growth_devs_cohort}{MFCLBiol}(object)

growth_devs_cohort(object, ...) <- value

\S4method{growth_devs_cohort}{MFCLBiol,FLCohort}(object) <- value

season_growth_pars(object, ...)

\S4method{season_growth_pars}{MFCLBiol}(object)

season_growth_pars(object, ...) <- value

\S4method{season_growth_pars}{MFCLBiol,numeric}(object) <- value

len_bias_pars(object, ...)

\S4method{len_bias_pars}{MFCLBiol}(object)

len_bias_pars(object, ...) <- value

\S4method{len_bias_pars}{MFCLBiol,numeric}(object) <- value

common_len_bias_pars(object, ...)

\S4method{common_len_bias_pars}{MFCLBiol}(object)

common_len_bias_pars(object, ...) <- value

\S4method{common_len_bias_pars}{MFCLBiol,numeric}(object) <- value

common_len_bias_coffs(object, ...)

\S4method{common_len_bias_coffs}{MFCLBiol}(object)

common_len_bias_coffs(object, ...) <- value

\S4method{common_len_bias_coffs}{MFCLBiol,numeric}(object) <- value

\S4method{dimensions}{MFCLBiol}(object)

\S4method{dimensions}{MFCLBiol,numeric}(object) <- value

\S4method{range}{MFCLBiol}(x)

\S4method{range}{MFCLBiol,missing,numeric}(x) <- value

flags(object, ...)

\S4method{flags}{MFCLFlags}(object)

flags(object, ...) <- value

\S4method{flags}{MFCLFlags,data.frame}(object) <- value

unused(object, ...)

\S4method{unused}{MFCLFlags}(object)

unused(object, ...) <- value

\S4method{unused}{MFCLFlags,list}(object) <- value

tag_shed_rate(object, ...)

\S4method{tag_shed_rate}{MFCLTagRep}(object)

tag_shed_rate(object, ...) <- value

\S4method{tag_shed_rate}{MFCLTagRep,numeric}(object) <- value

tag_fish_rep_rate(object, ...)

\S4method{tag_fish_rep_rate}{MFCLTagRep}(object)

tag_fish_rep_rate(object, ...) <- value

\S4method{tag_fish_rep_rate}{MFCLTagRep,array}(object) <- value

tag_fish_rep_grp(object, ...)

\S4method{tag_fish_rep_grp}{MFCLTagRep}(object)

tag_fish_rep_grp(object, ...) <- value

\S4method{tag_fish_rep_grp}{MFCLTagRep,array}(object) <- value

tag_fish_rep_flags(object, ...)

\S4method{tag_fish_rep_flags}{MFCLTagRep}(object)

tag_fish_rep_flags(object, ...) <- value

\S4method{tag_fish_rep_flags}{MFCLTagRep,array}(object) <- value

tag_fish_rep_target(object, ...)

\S4method{tag_fish_rep_target}{MFCLTagRep}(object)

tag_fish_rep_target(object, ...) <- value

\S4method{tag_fish_rep_target}{MFCLTagRep,array}(object) <- value

tag_fish_rep_pen(object, ...)

\S4method{tag_fish_rep_pen}{MFCLTagRep}(object)

tag_fish_rep_pen(object, ...) <- value

\S4method{tag_fish_rep_pen}{MFCLTagRep,array}(object) <- value

rep_rate_dev_coffs(object, ...)

\S4method{rep_rate_dev_coffs}{MFCLTagRep}(object)

rep_rate_dev_coffs(object, ...) <- value

\S4method{rep_rate_dev_coffs}{MFCLTagRep,list}(object) <- value

\S4method{range}{MFCLTagRep}(x)

\S4method{range}{MFCLTagRep,missing,numeric}(x) <- value

rec_init_pop_diff(object, ...)

\S4method{rec_init_pop_diff}{MFCLRec}(object)

rec_init_pop_diff(object, ...) <- value

\S4method{rec_init_pop_diff}{MFCLRec,numeric}(object) <- value

rec_times(object, ...)

\S4method{rec_times}{MFCLRec}(object)

rec_times(object, ...) <- value

\S4method{rec_times}{MFCLRec,numeric}(object) <- value

rel_rec(object, ...)

\S4method{rel_rec}{MFCLRec}(object)

rel_rec(object, ...) <- value

\S4method{rel_rec}{MFCLRec,FLQuant}(object) <- value

tot_pop(object, ...)

\S4method{tot_pop}{MFCLRec}(object)

tot_pop(object, ...) <- value

\S4method{tot_pop}{MFCLRec,numeric}(object) <- value

tot_pop_implicit(object, ...)

\S4method{tot_pop_implicit}{MFCLRec}(object)

tot_pop_implicit(object, ...) <- value

\S4method{tot_pop_implicit}{MFCLRec,numeric}(object) <- value

rel_ini_pop(object, ...)

\S4method{rel_ini_pop}{MFCLRec}(object)

rel_ini_pop(object, ...) <- value

\S4method{rel_ini_pop}{MFCLRec,array}(object) <- value

orth_coffs(object, ...)

\S4method{orth_coffs}{MFCLRec}(object)

orth_coffs(object, ...) <- value

\S4method{orth_coffs}{MFCLRec,matrix}(object) <- value

new_orth_coffs(object, ...)

\S4method{new_orth_coffs}{MFCLRec}(object)

new_orth_coffs(object, ...) <- value

\S4method{new_orth_coffs}{MFCLRec,numeric}(object) <- value

rec_orthogonal(object, ...)

\S4method{rec_orthogonal}{MFCLRec}(object)

rec_orthogonal(object, ...) <- value

\S4method{rec_orthogonal}{MFCLRec,FLQuant}(object) <- value

rec_standard_dim(object, ...)

\S4method{rec_standard_dim}{MFCLRec}(object)

rec_standard_dim(object, ...) <- value

\S4method{rec_standard_dim}{MFCLRec,numeric}(object) <- value

rec_standard(object, ...)

\S4method{rec_standard}{MFCLRec}(object)

rec_standard(object, ...) <- value

\S4method{rec_standard}{MFCLRec,FLQuant}(object) <- value

annual_rel_rec_coffs(object, ...)

\S4method{annual_rel_rec_coffs}{MFCLRec}(object)

annual_rel_rec_coffs(object, ...) <- value

\S4method{annual_rel_rec_coffs}{MFCLRec,matrix}(object) <- value

\S4method{range}{MFCLRec}(x)

\S4method{range}{MFCLRec,missing,numeric}(x) <- value

control_flags(object, ...)

\S4method{control_flags}{MFCLRegion}(object)

control_flags(object, ...) <- value

\S4method{control_flags}{MFCLRegion,matrix}(object) <- value

move_map(object, ...)

\S4method{move_map}{MFCLRegion}(object)

move_map(object, ...) <- value

\S4method{move_map}{MFCLRegion,numeric}(object) <- value

diff_coffs(object, ...)

\S4method{diff_coffs}{MFCLRegion}(object)

diff_coffs(object, ...) <- value

\S4method{diff_coffs}{MFCLRegion,matrix}(object) <- value

xdiff_coffs(object, ...)

\S4method{xdiff_coffs}{MFCLRegion}(object)

xdiff_coffs(object, ...) <- value

\S4method{xdiff_coffs}{MFCLRegion,matrix}(object) <- value

y1diff_coffs(object, ...)

\S4method{y1diff_coffs}{MFCLRegion}(object)

y1diff_coffs(object, ...) <- value

\S4method{y1diff_coffs}{MFCLRegion,matrix}(object) <- value

y2diff_coffs(object, ...)

\S4method{y2diff_coffs}{MFCLRegion}(object)

y2diff_coffs(object, ...) <- value

\S4method{y2diff_coffs}{MFCLRegion,matrix}(object) <- value

zdiff_coffs(object, ...)

\S4method{zdiff_coffs}{MFCLRegion}(object)

zdiff_coffs(object, ...) <- value

\S4method{zdiff_coffs}{MFCLRegion,matrix}(object) <- value

diff_coffs_mat(object, ...)

\S4method{diff_coffs_mat}{MFCLRegion}(object)

diff_coffs_mat(object, ...) <- value

\S4method{diff_coffs_mat}{MFCLRegion,matrix}(object) <- value

diff_coffs_age_ssn(object, ...)

\S4method{diff_coffs_age_ssn}{MFCLRegion}(object)

diff_coffs_age_ssn(object, ...) <- value

\S4method{diff_coffs_age_ssn}{MFCLRegion,array}(object) <- value

diff_coffs_age_period(object, ...)

\S4method{diff_coffs_age_period}{MFCLRegion}(object)

diff_coffs_age_period(object, ...) <- value

\S4method{diff_coffs_age_period}{MFCLRegion,array}(object) <- value

diff_coffs_age(object, ...)

\S4method{diff_coffs_age}{MFCLRegion}(object)

diff_coffs_age(object, ...) <- value

\S4method{diff_coffs_age}{MFCLRegion,array}(object) <- value

diff_coffs_nl(object, ...)

\S4method{diff_coffs_nl}{MFCLRegion}(object)

diff_coffs_nl(object, ...) <- value

\S4method{diff_coffs_nl}{MFCLRegion,array}(object) <- value

diff_coffs_priors(object, ...)

\S4method{diff_coffs_priors}{MFCLRegion}(object)

diff_coffs_priors(object, ...) <- value

\S4method{diff_coffs_priors}{MFCLRegion,array}(object) <- value

diff_coffs_age_priors(object, ...)

\S4method{diff_coffs_age_priors}{MFCLRegion}(object)

diff_coffs_age_priors(object, ...) <- value

\S4method{diff_coffs_age_priors}{MFCLRegion,array}(object) <- value

diff_coffs_nl_priors(object, ...)

\S4method{diff_coffs_nl_priors}{MFCLRegion}(object)

diff_coffs_nl_priors(object, ...) <- value

\S4method{diff_coffs_nl_priors}{MFCLRegion,array}(object) <- value

region_rec_var(object, ...)

\S4method{region_rec_var}{MFCLRegion}(object)

region_rec_var(object, ...) <- value

\S4method{region_rec_var}{MFCLRegion,FLQuant}(object) <- value

region_pars(object, ...)

\S4method{region_pars}{MFCLRegion}(object)

region_pars(object, ...) <- value

\S4method{region_pars}{MFCLRegion,matrix}(object) <- value

\S4method{range}{MFCLRegion}(x)

\S4method{range}{MFCLRegion,missing,numeric}(x) <- value

availability_coffs(object, ...)

\S4method{availability_coffs}{MFCLSel}(object)

availability_coffs(object, ...) <- value

\S4method{availability_coffs}{MFCLSel,FLQuant}(object) <- value

fishery_sel(object, ...)

\S4method{fishery_sel}{MFCLSel}(object)

fishery_sel(object, ...) <- value

\S4method{fishery_sel}{MFCLSel,FLQuant}(object) <- value

fishery_sel_age_comp(object, ...)

\S4method{fishery_sel_age_comp}{MFCLSel}(object)

fishery_sel_age_comp(object, ...) <- value

\S4method{fishery_sel_age_comp}{MFCLSel,FLQuant}(object) <- value

av_q_coffs(object, ...)

\S4method{av_q_coffs}{MFCLSel}(object)

av_q_coffs(object, ...) <- value

\S4method{av_q_coffs}{MFCLSel,FLQuant}(object) <- value

ini_q_coffs(object, ...)

\S4method{ini_q_coffs}{MFCLSel}(object)

ini_q_coffs(object, ...) <- value

\S4method{ini_q_coffs}{MFCLSel,FLQuant}(object) <- value

q0_miss(object, ...)

\S4method{q0_miss}{MFCLSel}(object)

q0_miss(object, ...) <- value

\S4method{q0_miss}{MFCLSel,FLQuant}(object) <- value

q_dev_coffs(object, ...)

\S4method{q_dev_coffs}{MFCLSel}(object)

q_dev_coffs(object, ...) <- value

\S4method{q_dev_coffs}{MFCLSel,list}(object) <- value

effort_dev_coffs(object, frq, ...)

\S4method{effort_dev_coffs}{MFCLSel,missing}(object)

effort_dev_coffs(object, ...) <- value

\S4method{effort_dev_coffs}{MFCLSel,list}(object) <- value

catch_dev_coffs(object, ...)

\S4method{catch_dev_coffs}{MFCLSel}(object)

catch_dev_coffs(object, ...) <- value

\S4method{catch_dev_coffs}{MFCLSel,list}(object) <- value

catch_dev_coffs_flag(object, ...)

\S4method{catch_dev_coffs_flag}{MFCLSel}(object)

catch_dev_coffs_flag(object, ...) <- value

\S4method{catch_dev_coffs_flag}{MFCLSel,numeric}(object) <- value

sel_dev_corr(object, ...)

\S4method{sel_dev_corr}{MFCLSel}(object)

sel_dev_corr(object, ...) <- value

\S4method{sel_dev_corr}{MFCLSel,FLQuant}(object) <- value

sel_dev_coffs(object, ...)

\S4method{sel_dev_coffs}{MFCLSel}(object)

sel_dev_coffs(object, ...) <- value

\S4method{sel_dev_coffs}{MFCLSel,matrix}(object) <- value

sel_dev_coffs2(object, ...)

\S4method{sel_dev_coffs2}{MFCLSel}(object)

sel_dev_coffs2(object, ...) <- value

\S4method{sel_dev_coffs2}{MFCLSel,list}(object) <- value

season_q_pars(object, ...)

\S4method{season_q_pars}{MFCLSel}(object)

season_q_pars(object, ...) <- value

\S4method{season_q_pars}{MFCLSel,matrix}(object) <- value

fish_params(object, ...)

\S4method{fish_params}{MFCLSel}(object)

fish_params(object, ...) <- value

\S4method{fish_params}{MFCLSel,matrix}(object) <- value

spp_params(object, ...)

\S4method{spp_params}{MFCLSel}(object)

spp_params(object, ...) <- value

\S4method{spp_params}{MFCLSel,matrix}(object) <- value

\S4method{range}{MFCLSel}(x)

\S4method{range}{MFCLSel,missing,numeric}(x) <- value

fm_level_devs(object, ...)

\S4method{fm_level_devs}{MFCLParBits}(object)

fm_level_devs(object, ...) <- value

\S4method{fm_level_devs}{MFCLParBits,character}(object) <- value

fm_level_regression_pars(object, ...)

\S4method{fm_level_regression_pars}{MFCLParBits}(object)

fm_level_regression_pars(object, ...) <- value

\S4method{fm_level_regression_pars}{MFCLParBits,matrix}(object) <- value

obj_fun(object, ...)

\S4method{obj_fun}{MFCLParBits}(object)

obj_fun(object, ...) <- value

\S4method{obj_fun}{MFCLParBits,numeric}(object) <- value

n_pars(object, ...)

\S4method{n_pars}{MFCLParBits}(object)

n_pars(object, ...) <- value

\S4method{n_pars}{MFCLParBits,numeric}(object) <- value

tag_lik(object, ...)

\S4method{tag_lik}{MFCLParBits}(object)

tag_lik(object, ...) <- value

\S4method{tag_lik}{MFCLParBits,numeric}(object) <- value

mn_len_pen(object, ...)

\S4method{mn_len_pen}{MFCLParBits}(object)

mn_len_pen(object, ...) <- value

\S4method{mn_len_pen}{MFCLParBits,numeric}(object) <- value

max_grad(object, ...)

\S4method{max_grad}{MFCLParBits}(object)

max_grad(object, ...) <- value

\S4method{max_grad}{MFCLParBits,numeric}(object) <- value

av_fish_mort_inst(object, ...)

\S4method{av_fish_mort_inst}{MFCLParBits}(object)

av_fish_mort_inst(object, ...) <- value

\S4method{av_fish_mort_inst}{MFCLParBits,numeric}(object) <- value

av_fish_mort_year(object, ...)

\S4method{av_fish_mort_year}{MFCLParBits}(object)

av_fish_mort_year(object, ...) <- value

\S4method{av_fish_mort_year}{MFCLParBits,numeric}(object) <- value

av_fish_mort_age(object, ...)

\S4method{av_fish_mort_age}{MFCLParBits}(object)

av_fish_mort_age(object, ...) <- value

\S4method{av_fish_mort_age}{MFCLParBits,numeric}(object) <- value

logistic_normal_params(object, ...)

\S4method{logistic_normal_params}{MFCLParBits}(object)

logistic_normal_params(object, ...) <- value

\S4method{logistic_normal_params}{MFCLParBits,character}(object) <- value

lagrangian(object, ...)

\S4method{lagrangian}{MFCLParBits}(object)

lagrangian(object, ...) <- value

\S4method{lagrangian}{MFCLParBits,character}(object) <- value

kludged_eq_coffs(object, ...)

\S4method{kludged_eq_coffs}{MFCLParBits}(object)

kludged_eq_coffs(object, ...) <- value

\S4method{kludged_eq_coffs}{MFCLParBits,array}(object) <- value

kludged_eq_level_coffs(object, ...)

\S4method{kludged_eq_level_coffs}{MFCLParBits}(object)

kludged_eq_level_coffs(object, ...) <- value

\S4method{kludged_eq_level_coffs}{MFCLParBits,numeric}(object) <- value

first_year(object, ...)

\S4method{first_year}{MFCLParBits}(object)

first_year(object, ...) <- value

\S4method{first_year}{MFCLParBits,numeric}(object) <- value

\S4method{range}{MFCLParBits}(x)

\S4method{range}{MFCLParBits,missing,numeric}(x) <- value

age_pars(object, ...)

\S4method{age_pars}{MFCLIniBits}(object)

age_pars(object, ...) <- value

\S4method{age_pars}{MFCLIniBits,matrix}(object) <- value

rec_dist(object, ...)

\S4method{rec_dist}{MFCLIniBits}(object)

rec_dist(object, ...) <- value

\S4method{rec_dist}{MFCLIniBits,numeric}(object) <- value

lw_params(object, ...)

\S4method{lw_params}{MFCLIni}(object)

lw_params(object, ...) <- value

\S4method{lw_params}{MFCLIniBits,numeric}(object) <- value

\S4method{sv}{MFCLIniBits,missing}(x)

sv(object, ...) <- value

\S4method{sv}{MFCLIniBits,numeric}(object) <- value

sd_length_at_age(object, ...)

\S4method{sd_length_at_age}{MFCLIniBits}(object)

sd_length_at_age(object, ...) <- value

\S4method{sd_length_at_age}{MFCLIniBits,numeric}(object) <- value

sd_length_dep(object, ...)

\S4method{sd_length_dep}{MFCLIniBits}(object)

sd_length_dep(object, ...) <- value

\S4method{sd_length_dep}{MFCLIniBits,numeric}(object) <- value

\S4method{tot_pop}{MFCLIniBits}(object)

\S4method{tot_pop}{MFCLIniBits,numeric}(object) <- value

tag_flags(object, ...)

\S4method{tag_flags}{MFCLIniBits}(object)

tag_flags(object, ...) <- value

\S4method{tag_flags}{MFCLIniBits,matrix}(object) <- value

fishery_realizations(object, ...)

\S4method{fishery_realizations}{MFCLRep}(object)

fishery_realizations(object, ...) <- value

\S4method{fishery_realizations}{MFCLRep,FLQuant}(object) <- value

mean_laa(object, ...)

\S4method{mean_laa}{MFCLRep}(object)

mean_laa(object, ...) <- value

\S4method{mean_laa}{MFCLRep,FLQuant}(object) <- value

mean_waa(object, ...)

\S4method{mean_waa}{MFCLRep}(object)

mean_waa(object, ...) <- value

\S4method{mean_waa}{MFCLRep,FLQuant}(object) <- value

sd_laa(object, ...)

\S4method{sd_laa}{MFCLRep}(object)

sd_laa(object, ...) <- value

\S4method{sd_laa}{MFCLRep,FLQuant}(object) <- value

m_at_age(object, ...)

\S4method{m_at_age}{MFCLRep}(object)

m_at_age(object, ...) <- value

\S4method{m_at_age}{MFCLRep,FLQuant}(object) <- value

sel(object, ...)

\S4method{sel}{MFCLRep}(object)

sel(object, ...) <- value

\S4method{sel}{MFCLRep,FLQuant}(object) <- value

q_fishery(object, ...)

\S4method{q_fishery}{MFCLRep}(object)

q_fishery(object, ...) <- value

\S4method{q_fishery}{MFCLRep,FLQuant}(object) <- value

q_effdev(object, ...)

\S4method{q_effdev}{MFCLRep}(object)

q_effdev(object, ...) <- value

\S4method{q_effdev}{MFCLRep,FLQuant}(object) <- value

fm(object, ...)

\S4method{fm}{MFCLRep}(object)

fm(object, ...) <- value

\S4method{fm}{MFCLRep,FLQuant}(object) <- value

fm_aggregated(object, ...)

\S4method{fm_aggregated}{MFCLRep}(object)

fm_aggregated(object, ...) <- value

\S4method{fm_aggregated}{MFCLRep,FLQuant}(object) <- value

popN(object, ...)

\S4method{popN}{MFCLRep}(object)

popN(object, ...) <- value

\S4method{popN}{MFCLRep,FLQuant}(object) <- value

rec_region(object, ...)

\S4method{rec_region}{MFCLRep}(object)

rec_region(object, ...) <- value

\S4method{rec_region}{MFCLRep,FLQuant}(object) <- value

totalBiomass(object, ...)

\S4method{totalBiomass}{MFCLRep}(object)

totalBiomass(object, ...) <- value

\S4method{totalBiomass}{MFCLRep,FLQuant}(object) <- value

totalBiomass_nofish(object, ...)

\S4method{totalBiomass_nofish}{MFCLRep}(object)

totalBiomass_nofish(object, ...) <- value

\S4method{totalBiomass_nofish}{MFCLRep,FLQuant}(object) <- value

adultBiomass(object, ...)

\S4method{adultBiomass}{MFCLRep}(object)

adultBiomass(object, ...) <- value

\S4method{adultBiomass}{MFCLRep,FLQuant}(object) <- value

adultBiomass_nofish(object, ...)

\S4method{adultBiomass_nofish}{MFCLRep}(object)

adultBiomass_nofish(object, ...) <- value

\S4method{adultBiomass_nofish}{MFCLRep,FLQuant}(object) <- value

vulnBiomass(object, ...)

\S4method{vulnBiomass}{MFCLRep}(object)

vulnBiomass(object, ...) <- value

\S4method{vulnBiomass}{MFCLRep,FLQuant}(object) <- value

srr(object, ...)

\S4method{srr}{MFCLRep}(object)

srr(object, ...) <- value

\S4method{srr}{MFCLRep,FLPar}(object) <- value

eq_ssb(object, ...)

\S4method{eq_ssb}{MFCLRep}(object)

eq_ssb(object, ...) <- value

\S4method{eq_ssb}{MFCLRep,FLQuant}(object) <- value

eq_ssb_obs(object, ...)

\S4method{eq_ssb_obs}{MFCLRep}(object)

eq_ssb_obs(object, ...) <- value

\S4method{eq_ssb_obs}{MFCLRep,FLQuant}(object) <- value

eq_rec(object, ...)

\S4method{eq_rec}{MFCLRep}(object)

eq_rec(object, ...) <- value

\S4method{eq_rec}{MFCLRep,FLQuant}(object) <- value

eq_rec_obs(object, ...)

\S4method{eq_rec_obs}{MFCLRep}(object)

eq_rec_obs(object, ...) <- value

\S4method{eq_rec_obs}{MFCLRep,FLQuant}(object) <- value

catch_obs(object, ...)

\S4method{catch_obs}{MFCLRep}(object)

catch_obs(object, ...) <- value

\S4method{catch_obs}{MFCLRep,FLQuant}(object) <- value

catch_pred(object, ...)

\S4method{catch_pred}{MFCLRep}(object)

catch_pred(object, ...) <- value

\S4method{catch_pred}{MFCLRep,FLQuant}(object) <- value

cpue_obs(object, ...)

\S4method{cpue_obs}{MFCLRep}(object)

cpue_obs(object, ...) <- value

\S4method{cpue_obs}{MFCLRep,FLQuant}(object) <- value

cpue_pred(object, ...)

\S4method{cpue_pred}{MFCLRep}(object)

cpue_pred(object, ...) <- value

\S4method{cpue_pred}{MFCLRep,FLQuant}(object) <- value

eq_biomass(object, ...)

\S4method{eq_biomass}{MFCLRep}(object)

eq_biomass(object, ...) <- value

\S4method{eq_biomass}{MFCLRep,FLQuant}(object) <- value

eq_yield(object, ...)

\S4method{eq_yield}{MFCLRep}(object)

eq_yield(object, ...) <- value

\S4method{eq_yield}{MFCLRep,FLQuant}(object) <- value

MSY(object, ...)

\S4method{MSY}{MFCLRep}(object)

MSY(object, ...) <- value

\S4method{MSY}{MFCLRep,numeric}(object) <- value

FMSY(object, ...)

\S4method{FMSY}{MFCLRep}(object)

FMSY(object, ...) <- value

\S4method{FMSY}{MFCLRep,numeric}(object) <- value

BMSY(object, ...)

\S4method{BMSY}{MFCLRep}(object)

BMSY(object, ...) <- value

\S4method{BMSY}{MFCLRep,numeric}(object) <- value

FFMSY_ts(object, ...)

\S4method{FFMSY_ts}{MFCLRep}(object)

FFMSY_ts(object, ...) <- value

\S4method{FFMSY_ts}{MFCLRep,FLQuant}(object) <- value

ABBMSY_ts(object, ...)

\S4method{ABBMSY_ts}{MFCLRep}(object)

ABBMSY_ts(object, ...) <- value

\S4method{ABBMSY_ts}{MFCLRep,FLQuant}(object) <- value

ABBMSY(object, ...)

\S4method{ABBMSY}{MFCLRep}(object)

ABBMSY(object, ...) <- value

\S4method{ABBMSY}{MFCLRep,numeric}(object) <- value

TBBMSY(object, ...)

\S4method{TBBMSY}{MFCLRep}(object)

TBBMSY(object, ...) <- value

\S4method{TBBMSY}{MFCLRep,numeric}(object) <- value

Fmult(object, ...)

\S4method{Fmult}{MFCLRep}(object)

Fmult(object, ...) <- value

\S4method{Fmult}{MFCLRep,numeric}(object) <- value

AggregateF(object, ...)

\S4method{AggregateF}{MFCLRep}(object)

AggregateF(object, ...) <- value

\S4method{AggregateF}{MFCLRep,FLQuant}(object) <- value

\S4method{dimensions}{MFCLRep}(object)

\S4method{dimensions}{MFCLRep,numeric}(object) <- value

\S4method{range}{MFCLRep}(x)

\S4method{range}{MFCLRep,missing,numeric}(x) <- value

nyears(object, ...)

\S4method{nyears}{MFCLprojControl}(object)

nyears(object, ...) <- value

\S4method{nyears}{MFCLprojControl,numeric}(object) <- value

nsims(object, ...)

\S4method{nsims}{MFCLprojControl}(object)

nsims(object, ...) <- value

\S4method{nsims}{MFCLprojControl,numeric}(object) <- value

avyrs(object, ...)

\S4method{avyrs}{MFCLprojControl}(object)

avyrs(object, ...) <- value

\S4method{avyrs}{MFCLprojControl,character}(object) <- value

fprojyr(object, ...)

\S4method{fprojyr}{MFCLprojControl}(object)

fprojyr(object, ...) <- value

\S4method{fprojyr}{MFCLprojControl,numeric}(object) <- value

controls(object, ...)

\S4method{controls}{MFCLprojControl}(object)

controls(object, ...) <- value

\S4method{controls}{MFCLprojControl,data.frame}(object) <- value

caeff(object, ...)

\S4method{caeff}{MFCLprojControl}(object)

caeff(object, ...) <- value

\S4method{caeff}{MFCLprojControl,numeric}(object) <- value

scaler(object, ...)

\S4method{scaler}{MFCLprojControl}(object)

scaler(object, ...) <- value

\S4method{scaler}{MFCLprojControl,numeric}(object) <- value

ess(object, ...)

\S4method{ess}{MFCLprojControl}(object)

ess(object, ...) <- value

\S4method{ess}{MFCLprojControl,data.frame}(object) <- value

\S4method{range}{MFCLFrq}(x)

\S4method{range}{MFCLFrq,missing,numeric}(x) <- value

\S4method{range}{MFCLPar}(x)

\S4method{range}{MFCLPar,missing,numeric}(x) <- value

\S4method{dimensions}{MFCLPar}(object)

\S4method{dimensions}{MFCLPar,numeric}(object) <- value

\S4method{range}{MFCLIni}(x)

\S4method{range}{MFCLIni,missing,numeric}(x) <- value

ini_version(object, ...)

\S4method{ini_version}{MFCLIniBits}(object)

ini_version(object, ...) <- value

\S4method{ini_version}{MFCLIniBits,numeric}(object) <- value

region_flags(object, ...)

\S4method{region_flags}{MFCLIniBits}(object)

region_flags(object, ...) <- value

\S4method{region_flags}{MFCLIniBits,matrix}(object) <- value

\S4method{range}{MFCLRep}(x)

\S4method{range}{MFCLRep,missing,numeric}(x) <- value

\S4method{dimensions}{MFCLRep}(object)

\S4method{dimensions}{MFCLRep,numeric}(object) <- value

release_groups_proj(object, ...)

\S4method{release_groups_proj}{MFCLTagProj}(object)

release_groups_proj(object, ...) <- value

\S4method{release_groups_proj}{MFCLTagProj,numeric}(object) <- value

releases_proj(object, ...)

\S4method{releases_proj}{MFCLTagProj}(object)

releases_proj(object, ...) <- value

\S4method{releases_proj}{MFCLTagProj,data.frame}(object) <- value

rep_rate_proj(object, ...)

\S4method{rep_rate_proj}{MFCLTagProj}(object)

rep_rate_proj(object, ...) <- value

\S4method{rep_rate_proj}{MFCLTagProj,array}(object) <- value

\S4method{range}{MFCLTagProj}(x)

\S4method{range}{MFCLTagProj,missing,numeric}(x) <- value

release_groups(object, ...)

\S4method{release_groups}{MFCLTag}(object)

release_groups(object, ...) <- value

\S4method{release_groups}{MFCLTag,numeric}(object) <- value

release_lengths(object, ...)

\S4method{release_lengths}{MFCLTag}(object)

release_lengths(object, ...) <- value

\S4method{release_lengths}{MFCLTag,numeric}(object) <- value

recoveries(object, ...)

\S4method{recoveries}{MFCLTag}(object)

recoveries(object, ...) <- value

\S4method{recoveries}{MFCLTag,numeric}(object) <- value

releases(object, ...)

\S4method{releases}{MFCLTag}(object)

releases(object, ...) <- value

\S4method{releases}{MFCLTag,data.frame}(object) <- value

recaptures(object, ...)

\S4method{recaptures}{MFCLTag}(object)

recaptures(object, ...) <- value

\S4method{recaptures}{MFCLTag,data.frame}(object) <- value

\S4method{range}{MFCLTag}(x)

\S4method{range}{MFCLTag,missing,numeric}(x) <- value

total_catch(object, ...)

\S4method{total_catch}{MFCLCatch}(object)

total_catch(object, ...) <- value

\S4method{total_catch}{MFCLCatch,FLQuant}(object) <- value

fishery_catch(object, ...)

\S4method{fishery_catch}{MFCLCatch}(object)

fishery_catch(object, ...) <- value

\S4method{fishery_catch}{MFCLCatch,FLQuant}(object) <- value

\S4method{range}{MFCLCatch}(x)

\S4method{range}{MFCLCatch,missing,numeric}(x) <- value

laa(object, ...)

\S4method{laa}{MFCLLenFit}(object)

laa(object, ...) <- value

\S4method{laa}{MFCLLenFit,FLQuant}(object) <- value

lenfits(object, ...)

\S4method{lenfits}{MFCLLenFit}(object)

lenfits(object, ...) <- value

\S4method{lenfits}{MFCLLenFit,data.frame}(object) <- value

lenagefits(object, ...)

\S4method{lenagefits}{MFCLLenFit}(object)

lenagefits(object, ...) <- value

\S4method{lenagefits}{MFCLLenFit,data.frame}(object) <- value

\S4method{range}{MFCLLenFit}(x)

\S4method{range}{MFCLLenFit,missing,numeric}(x) <- value

waa(object, ...)

\S4method{waa}{MFCLWgtFit}(object)

waa(object, ...) <- value

\S4method{waa}{MFCLWgtFit,FLQuant}(object) <- value

wgtfits(object, ...)

\S4method{wgtfits}{MFCLWgtFit}(object)

wgtfits(object, ...) <- value

\S4method{wgtfits}{MFCLWgtFit,data.frame}(object) <- value

wgtagefits(object, ...)

\S4method{wgtagefits}{MFCLWgtFit}(object)

wgtagefits(object, ...) <- value

\S4method{wgtagefits}{MFCLWgtFit,data.frame}(object) <- value

\S4method{range}{MFCLWgtFit}(x)

\S4method{range}{MFCLWgtFit,missing,numeric}(x) <- value

bh_steep_contrib(object, ...)

\S4method{bh_steep_contrib}{MFCLLikelihood}(object)

bh_steep_contrib(object, ...) <- value

\S4method{bh_steep_contrib}{MFCLLikelihood,numeric}(object) <- value

effort_dev_penalty(object, ...)

\S4method{effort_dev_penalty}{MFCLLikelihood}(object)

effort_dev_penalty(object, ...) <- value

\S4method{effort_dev_penalty}{MFCLLikelihood,numeric}(object) <- value

q_dev_pen_fish(object, ...)

\S4method{q_dev_pen_fish}{MFCLLikelihood}(object)

q_dev_pen_fish(object, ...) <- value

\S4method{q_dev_pen_fish}{MFCLLikelihood,numeric}(object) <- value

q_dev_pen_fish_grp(object, ...)

\S4method{q_dev_pen_fish_grp}{MFCLLikelihood}(object)

q_dev_pen_fish_grp(object, ...) <- value

\S4method{q_dev_pen_fish_grp}{MFCLLikelihood,numeric}(object) <- value

total_length_fish(object, ...)

\S4method{total_length_fish}{MFCLLikelihood}(object)

total_length_fish(object, ...) <- value

\S4method{total_length_fish}{MFCLLikelihood,numeric}(object) <- value

length_fish(object, ...)

\S4method{length_fish}{MFCLLikelihood}(object)

length_fish(object, ...) <- value

\S4method{length_fish}{MFCLLikelihood,list}(object) <- value

total_weight_fish(object, ...)

\S4method{total_weight_fish}{MFCLLikelihood}(object)

total_weight_fish(object, ...) <- value

\S4method{total_weight_fish}{MFCLLikelihood,numeric}(object) <- value

weight_fish(object, ...)

\S4method{weight_fish}{MFCLLikelihood}(object)

weight_fish(object, ...) <- value

\S4method{weight_fish}{MFCLLikelihood,list}(object) <- value

total_catch_fish(object, ...)

\S4method{total_catch_fish}{MFCLLikelihood}(object)

total_catch_fish(object, ...) <- value

\S4method{total_catch_fish}{MFCLLikelihood,numeric}(object) <- value

catch_fish(object, ...)

\S4method{catch_fish}{MFCLLikelihood}(object)

catch_fish(object, ...) <- value

\S4method{catch_fish}{MFCLLikelihood,list}(object) <- value

tag_rel_fish(object, ...)

\S4method{tag_rel_fish}{MFCLLikelihood}(object)

tag_rel_fish(object, ...) <- value

\S4method{tag_rel_fish}{MFCLLikelihood,list}(object) <- value

survey_index(object, ...)

\S4method{survey_index}{MFCLLikelihood}(object)

survey_index(object, ...) <- value

\S4method{survey_index}{MFCLLikelihood,numeric}(object) <- value

age_length(object, ...)

\S4method{age_length}{MFCLLikelihood}(object)

age_length(object, ...) <- value

\S4method{age_length}{MFCLLikelihood,numeric}(object) <- value

\S4method{dimensions}{MFCLLikelihood}(object)

\S4method{dimensions}{MFCLLikelihood,numeric}(object) <- value

catcheff(object, ...)

\S4method{catcheff}{MFCLPseudo}(object)

catcheff(object, ...) <- value

\S4method{catcheff}{MFCLPseudo,data.frame}(object) <- value

l_frq(object, ...)

\S4method{l_frq}{MFCLPseudo}(object)

l_frq(object, ...) <- value

\S4method{l_frq}{MFCLPseudo,data.frame}(object) <- value

w_frq(object, ...)

\S4method{w_frq}{MFCLPseudo}(object, ...)

w_frq(object, ...) <- value

\S4method{w_frq}{MFCLPseudo,data.frame}(object) <- value

catch_sd(object, ...)

\S4method{catch_sd}{MFCLPseudoControl}(object)

catch_sd(object, ...) <- value

\S4method{catch_sd}{MFCLPseudoControl,numeric}(object) <- value

effort_sd(object, ...)

\S4method{effort_sd}{MFCLPseudoControl}(object)

effort_sd(object, ...) <- value

\S4method{effort_sd}{MFCLPseudoControl,numeric}(object) <- value

tag_controls(object, ...)

\S4method{tag_controls}{MFCLPseudoControl}(object)

tag_controls(object, ...) <- value

\S4method{tag_controls}{MFCLPseudoControl,data.frame}(object) <- value

random_seeds(object, ...)

\S4method{random_seeds}{MFCLPseudoControl}(object)

random_seeds(object, ...) <- value

\S4method{random_seeds}{MFCLPseudoControl,numeric}(object) <- value

\S4method{tag_fish_rep_rate}{MFCLPseudoControl}(object)

\S4method{tag_fish_rep_rate}{MFCLPseudoControl,numeric}(object) <- value

doitall(object, ...)

\S4method{doitall}{MFCLEMControl}(object)

doitall(object, ...) <- value

\S4method{doitall}{MFCLEMControl,function}(object) <- value

\S4method{tag_fish_rep_rate}{MFCLEMControl}(object)

\S4method{tag_fish_rep_rate}{MFCLEMControl,numeric}(object) <- value

hcr(object, ...)

\S4method{hcr}{MFCLMSEControl}(object)

hcr(object, ...) <- value

\S4method{hcr}{MFCLMSEControl,character}(object) <- value

hcr_params(object, ...)

\S4method{hcr_params}{MFCLMSEControl}(object)

hcr_params(object, ...) <- value

\S4method{hcr_params}{MFCLMSEControl,numeric}(object) <- value

itn(object, ...)

\S4method{itn}{MFCLMSEControl}(object)

itn(object, ...) <- value

effort_creep(object, ...)

\S4method{effort_creep}{MFCLMSEControl}(object)

effort_creep(object, ...) <- value

\S4method{effort_creep}{MFCLMSEControl,numeric}(object) <- value

effort_creep_fish(object, ...)

\S4method{effort_creep_fish}{MFCLMSEControl}(object)

effort_creep_fish(object, ...) <- value

\S4method{effort_creep_fish}{MFCLMSEControl,numeric}(object) <- value

ALK(object, ...)

\S4method{ALK}{MFCLALK}(object)

ALK(object, ...) <- value

\S4method{ALK}{MFCLALK,data.frame}(object) <- value

ESS(object, ...)

\S4method{ESS}{MFCLALK}(object)

ESS(object, ...) <- value

\S4method{ESS}{MFCLALK,numeric}(object) <- value

\S4method{range}{MFCLALK}(x)

\S4method{range}{MFCLALK,missing,numeric}(x) <- value

Eq_calcs(object, ...)

\S4method{Eq_calcs}{MFCLEquilibrium}(object)

Eq_calcs(object, ...) <- value

\S4method{Eq_calcs}{MFCLEquilibrium,data.frame}(object) <- value

\S4method{index}{MFCLRegScale}(object)

\S4method{index}{MFCLRegScale,matrix}(object) <- value

\S4method{range}{MFCLRegScale}(x)

\S4method{range}{MFCLRegScale,missing,numeric}(x) <- value

\S4method{dimensions}{MFCLRegScale}(object)

\S4method{dimensions}{MFCLRegScale,numeric}(object) <- value
}
\arguments{
\item{object}{An object of type 'MFCLxxx', e.g. 'MFCLFrq'}

\item{...}{Other arguments}

\item{value}{Replacement value for the replacement methods}

\item{x}{An object of type 'MFCLxxx', e.g. 'MFCLFrq'}
}
\description{
A metric butt-tonne of accessor methods.
}
\details{
Accessors for extracting and writing data from / to an MFCLxxx object.
They are usually named after the slot names of the object class.
}
