% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdultJuvenileF.R
\name{AdultJuvenileF}
\alias{AdultJuvenileF}
\title{Adult and Juvenile F}
\usage{
AdultJuvenileF(rep, par)
}
\arguments{
\item{rep}{object of class MFCLRep.}

\item{par}{object of class MFCLPar.}
}
\value{
Data frame containing four columns:
\item{year}{year}
\item{area}{region, including \code{"all"}}
\item{stage}{stage, \code{"adult"} and \code{"juvenile"}}
\item{f}{fishing mortality rate, annual}
}
\description{
Calculate adult and juvenile fishing mortality.
}
\note{
Adult fishing mortality is calculated as the weighted average fishing
mortality, weighted by the maturity ogive.

Juvenile fishing mortality is calculated as the weighted average fishing
mortality, weighted by the inverse \code{(1-p)} of the maturity ogive.

In each year and region:
\preformatted{
f.adult <- weighted.mean(f.age, w=p.adult)  # correct
f.juven <- weighted.mean(f.age, w=p.juven)  # correct
}
}
\section{Warning}{

The obsolete \pkg{diags4MFCL} package provided a similar function called
\code{plot.F.temporal}, but that function had errors and should not be used.
It was based on wrong calculations of fishing mortality:
\preformatted{
.[,dead.juv:=f*juv] \%>\%              # wrong
.[,dead.adult:=f*adult] \%>\%          # wrong
.[,F.juv:=(dead.juv/juv)] \%>\%        # wrong
.[,F.adult:=(dead.adult/adult)] \%>\%  # wrong
}
The \code{AdultJuvenileF} function in the \pkg{FLR4MFCL} package is the
correct function to use for calculating adult and juvenile fishing
mortalities.
}

\examples{
\dontrun{
AdultJuvenileF(rep, par)
}

}
\seealso{
\code{\link{fm}} and \code{\link{fm_aggregated}} are used to access the
fishing mortalities.

\code{\link{mat}} is used to access the maturity ogive.

\code{\link{weighted.mean}} is the underlying function used to calculate the
weighted average fishing mortality.
}
