% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagTools.r
\name{ageflag}
\alias{ageflag}
\title{ageflag}
\usage{
ageflag(x, flags, ...)
}
\arguments{
\item{flags}{numeric vector of age flag number(s).}

\item{par}{object of class MFCLFlag()}
}
\value{
A data frame of sorted flag settings
}
\description{
A simple wrapper function for \code{\link{flagval}}
}
\details{
get age flag settings
}
\examples{
data(par)
ageflag(par, c(32,92))

}
\seealso{
\code{\link{flagval}} calls this function to retrieve flag settings.

\code{\link{read.MFCLFlags}} reads flag settings from a par file.
}
