% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuantMethods.r
\name{ats}
\alias{ats}
\title{ats}
\usage{
ats(quant)
}
\arguments{
\item{quant}{An FLQuant object.}
}
\value{
An object of class FLQuant.
}
\description{
Converts a quarterly time series object into a 6 dimensional, seasonally structured FLQuant (i.e. the opposite of qts).
}
