% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfcl.r
\name{availableMFCLversions}
\alias{availableMFCLversions}
\title{availableMFCLversions}
\usage{
availableMFCLversions()
}
\value{
Returns a list of available MFCL versions.
}
\description{
Returns the available versions of MFCL within the package.
Currently only the 2015 development version of MFCL is available.
}
\examples{
availableMFCLversions()

}
