% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mseTools.r
\name{catchWfromcatchN}
\alias{catchWfromcatchN}
\title{catchWfromcatchN}
\usage{
catchWfromcatchN(pseudo, projN, rep, ctrl, fsh, itn, ...)
}
\arguments{
\item{pseudo}{An object of class MFCLPseudo,}

\item{...}{Additional argument list that might not ever
be used.}

\item{projpar}{An object of class MFCLPar,}
}
\value{
Returns a data frame of the converted catch in weight.
}
\description{
Calculates the catch in weight from MFCL projected numbers at age and selection patterns.
Steps are:
1 - from the catcheff data frame you know what total catches should be (in number)
2 - use the selection params in the rep file to calculate vulnerable numbers
3 - use the selection and weight at age params in the rep file to calculate vulnerable biomass
4 - catch weight = vulB * (CatchN / vulN).
}
\examples{
#checkUnitDimnames(MFCLRep())
}
