% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.r
\name{checkUnitDimnames}
\alias{checkUnitDimnames}
\title{checkUnitDimnames}
\usage{
checkUnitDimnames(obj, nfisheries)
}
\arguments{
\item{obj}{An object of class MFCL eg. MFCLFrq, MFCLPar, etc.}

\item{nfisheries}{The number of fisheries in the model}

\item{...}{Additional argument list that might not ever be used.}
}
\value{
Returns an object of the same class but with FLQuant unit dimension
re-ordered as appropriate.
}
\description{
Re-orders the unit dimnames in ascending order instead
of default ordering of FLQuant method which uses the year dimension
}
\examples{
\dontrun{
checkUnitDimnames(MFCLRep())
}
}
