% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{corFilter}
\alias{corFilter}
\title{Correlation Filter}
\usage{
corFilter(x, level = 0.9, sort = TRUE)
}
\arguments{
\item{x}{correlations as a three-column data frame.}

\item{level}{filter level, a value from 0 to 1.}

\item{sort}{whether to sort the resulting data frame by absolute correlation.}
}
\value{
Data frame containing all correlations \code{|r| >= level}.
}
\description{
Filter correlations to show the highest correlations.
}
\examples{
data(cordf)
corFilter(cordf)
corFilter(cordf, level=0.99)
corFilter(cordf, level=0.99, sort=FALSE)

}
\seealso{
\code{\link{read.MFCLCor}} reads a correlation matrix from a file.

\code{\link{mat2MFCLCor}} converts a correlation matrix to a data frame.

\code{\link{corLabel}} labels correlations and produces a frequency table.
}
