% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{corLabel}
\alias{corLabel}
\title{Correlation Labels}
\usage{
corLabel(x)
}
\arguments{
\item{x}{correlations as a three-column data frame.}
}
\value{
Data frame with label counts.
}
\description{
Label correlations and produce a frequency table.
}
\examples{
data(cordf)
corLabel(cordf)

}
\seealso{
\code{\link{read.MFCLCor}} reads a correlation matrix from a file.

\code{\link{mat2MFCLCor}} converts a correlation matrix to a data frame.

\code{\link{corFilter}} filters correlations to show the highest
correlations.
}
