% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\docType{data}
\name{cormat}
\alias{cormat}
\title{Correlation Matrix}
\format{
Matrix with 5 rows and 5 columns, showing the correlation between parameter
pairs.
}
\usage{
cormat
}
\description{
Correlation matrix for examples.
}
\examples{
data(cormat)
cormat
mat2MFCLCor(cormat)
}
\seealso{
\code{\link{cordf}} is a three-column data frame of correlations.

\code{\link{mat2MFCLCor}} converts a correlation matrix to a data frame.
}
