% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagTools.r
\name{diffFlags}
\alias{diffFlags}
\alias{flagDiff}
\title{Diff Flags}
\usage{
diffFlags(par1, par2, all = TRUE, flaglist = NULL)

flagDiff(...)
}
\arguments{
\item{par1}{MFCL flags from model run 1.}

\item{par2}{MFCL flags from model run 2.}

\item{flaglist}{optional filename to use instead of the built-in
\file{flaglist.csv} lookup table.}

\item{\dots}{passed to \code{diffFlags}.}
}
\value{
A data frame showing flag settings where par1 and par2 are different, along
with a column showing the meaning of each flag.
}
\description{
Show differences in flag settings between two model runs.
}
\details{
The \code{par1} and \code{par2} objects can be any of the following:
\enumerate{
\item folder containing a par file
\item filename of a par file
\item \code{MFCLPar} object
\item \code{MFCLFlags} object
\item \code{data.frame} containing flag settings
}
}
\note{
\code{flagDiff} is an older name of this function. To support legacy scripts,
a call to the old function is simply forwarded to \code{diffFlags}.
}
\examples{
data(par)
par1 <- par2 <- par

# Different flag value
flags(par2)[20,"value"] <- 12
diffFlags(par1, par2)

# When flag is specified in par1 but not in par2
flags(par1) <- rbind(flags(par1), c(-10269, 1, 1))
diffFlags(par1, par2)             # default is to show par2 as NA
diffFlags(par1, par2, all=FALSE)  # all=FALSE omits such comparisons

}
\seealso{
This function calls \code{\link{flagMeaning}} to add the column showing the
meaning of each flag.

\code{\link{diffFlagsStepwise}} shows differences in flag settings between
stepwise model runs.

\code{\link{read.MFCLFlags}} reads flag settings from a par file.
}
