% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagTools.r
\name{diffFlagsStepwise}
\alias{diffFlagsStepwise}
\title{Diff Flags Stepwise}
\usage{
diffFlagsStepwise(stepdir, ...)
}
\arguments{
\item{stepdir}{is a directory containing model runs in subdirectories.}

\item{\dots}{passed to \code{diffFlags}.}
}
\value{
A list of data frames showing differences in flag settings between stepwise
model runs.
}
\description{
Show differences in flag settings between stepwise model runs.
}
\examples{
\dontrun{
diffFlagsStepwise("stepwise")

# Show number of changed flags in each step
lapply(diffFlagsStepwise("stepwise"), nrow)
}

}
\seealso{
\code{\link{diffFlags}} shows differences in flag settings between two model
runs.

\code{\link{read.MFCLFlags}} reads flag settings from a par file.
}
