% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniTools.R
\name{diffIniStepwise}
\alias{diffIniStepwise}
\title{Diff Ini Stepwise}
\usage{
diffIniStepwise(
  stepdir,
  models = dir(stepdir, full.names = TRUE),
  labels = basename(models),
  ...
)
}
\arguments{
\item{stepdir}{directory containing model runs in subdirectories.}

\item{models}{optional vector of directory names to manually specify stepwise
models to compare.}

\item{labels}{optional vector of short labels to describe the stepwise
models.}

\item{\dots}{passed to \code{diffFlags}.}
}
\value{
A data frame showing the differences in the \verb{ini} file between stepwise
model runs.
}
\description{
Show differences in the \verb{ini} file between stepwise model runs.
}
\details{
Generally, the user only needs to specify \code{stepdir}. If this top
directory contains stepwise model runs as subdirectories, then the default
values of \code{models} and \code{labels} will infer the correct paths and
model names.

If the stepwise model runs are not organized in a straightforward way, the
\code{models} and \code{labels} arguments can be passed explicitly.
}
\examples{
\dontrun{
yft_dir <- "//penguin/assessments/yft/2020_review/analysis/stepwise"
yft_diffs <- diffIniStepwise(yft_dir)
summary_ini(inifile)

# Unusual directory structure of BET 2020 stepwise models
bet_dir <- "//penguin/assessments/bet/2020/2020_stepwise"
bet_models <- file.path(dir(bet_dir, full.names=TRUE), "10N")
bet_labels <- dir(bet_dir)
bet_diffs <- diffIniStepwise(bet_dir, bet_models, bet_labels)
}

}
\seealso{
\code{\link{summary_ini}} summarizes the contents of an \verb{ini} file.

\code{\link{read.MFCLIni}} reads an \verb{ini} file.
}
