% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{finalPar}
\alias{finalPar}
\title{Final Par File}
\usage{
finalPar(folder, pattern = "^[0-9][0-9]\\\\.par$", full = TRUE, quiet = FALSE)
}
\arguments{
\item{folder}{model run directory name.}

\item{pattern}{pattern describing \code{par} filenames. The default pattern
is two digits, followed by \code{.par}.}

\item{full}{whether to return the \code{par} filename with the full path.}

\item{quiet}{whether to suppress the on-screen reporting of the final
\code{par} file found.}
}
\value{
The final \code{par} filename as a string.
}
\description{
Find the final \code{par} file in a model run directory.
}
\note{
If a folder contains \code{00.par}, \code{01.par}, \code{...}, \code{14.par},
then the final \code{par} file is \code{14.par}.
}
\examples{
\dontrun{
finalPar("modelrun")
}

}
\seealso{
\code{\link{finalRep}} finds the final \code{rep} file.
}
