% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{finalRep}
\alias{finalRep}
\title{Final Rep File}
\usage{
finalRep(
  folder,
  pattern = "^plot-[0-9][0-9]\\\\.par.rep$",
  full = TRUE,
  quiet = TRUE
)
}
\arguments{
\item{folder}{model run directory name.}

\item{pattern}{pattern describing \code{rep} filenames. The default pattern
is \code{plot-} and two digits, followed by \code{.rep}.}

\item{full}{whether to return the \code{rep} filename with the full path.}

\item{quiet}{whether to suppress the on-screen reporting of the final
\code{rep} file found.}
}
\value{
The final \code{rep} filename as a string.
}
\description{
Find the final \code{rep} file in a model run directory.
}
\note{
If a folder contains \code{plot-09.par.rep}, \code{plot-10.par.rep},
\code{...}, \code{plot-14.par.rep}, then the final \code{rep} file is
\code{plot-14.par.rep}.
}
\examples{
\dontrun{
finalRep("modelrun")
}

}
\seealso{
\code{\link{finalPar}} finds the final \code{par} file.
}
