% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{firstYear}
\alias{firstYear}
\title{First Year}
\usage{
firstYear(folder)
}
\arguments{
\item{folder}{is a folder containing at least one \verb{.rep} file.}
}
\value{
First year as an integer.
}
\description{
Read the first year in the model from the \verb{.rep} file.
}
\note{
Can be useful to find a suitable \code{first.yr} argument for
\code{read.MFCLPar}.
}
\examples{
\dontrun{
first.yr <- firstYear("modelrun")
parfile <- finalPar("modelrun")
par <- read.MFCLPar(parfile, first.yr)
}

}
\seealso{
\code{\link{read.MFCLPar}}.
}
