% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagTools.r
\name{fishflag}
\alias{fishflag}
\alias{fishflag,MFCLFlags-method}
\title{Get Fish Flag Settings}
\usage{
fishflag(x, flags, ...)

\S4method{fishflag}{MFCLFlags}(x, flags)
}
\arguments{
\item{x}{object of class \code{MFCLFlags}.}

\item{flags}{numeric vector of fish flag number(s).}

\item{\dots}{currently ignored.}
}
\value{
A data frame of sorted flag settings.
}
\description{
A simple wrapper function for \code{\link{flagval}}.
}
\examples{
data(par)
fishflag(par, c(1,15))

}
\seealso{
\code{\link{flagval}} calls this function to retrieve flag settings.

\code{\link{read.MFCLFlags}} reads flag settings from a \verb{par} file.
}
