% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagTools.r
\name{flagDiff}
\alias{flagDiff}
\title{flagDiff}
\usage{
flagDiff(par1, par2)
}
\arguments{
\item{par}{An object of class MFCLPar}
}
\value{
A data frame of flag settings for par1 and par2
}
\description{
flag differences between two par files
}
\examples{
data(par)
par1 <- par2 <- par
flags(par2)[20,"value"] <- 12
flagDiff(par1, par2)

}
