% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagTools.r
\name{flagMeaning}
\alias{flagMeaning}
\title{Flag Meaning}
\usage{
flagMeaning(flags, flaglist = NULL)
}
\arguments{
\item{flags}{MFCL flags from a model run.}

\item{flaglist}{optional filename to use instead of the built-in
\file{flaglist.csv} lookup table.}
}
\value{
A data frame with the same columns as \code{flags} plus a column called
\code{meaning}.
}
\description{
Show the meaning of flags, based on a lookup table.
}
\details{
The \code{flags} object can be any of the following:
\enumerate{
\item filename pointing to a par file
\item \code{MFCLPar} object
\item \code{MFCLFlags} object
\item \code{data.frame} containing flag settings
}
}
\examples{
data(par)
flagMeaning(par)

}
\seealso{
\code{\link{diffFlags}} calls this function to show the meaning of flags that
are different between two model runs.

\code{\link{read.MFCLFlags}} reads flag settings from a par file.
}
