% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagTools.r
\name{flagSort}
\alias{flagSort}
\title{Flag Sort}
\usage{
flagSort(flags)
}
\arguments{
\item{flags}{dataframe of MFCL flags from MFCLPar object.}
}
\value{
A data frame of sorted flag settings
}
\description{
Sort flags by flagtype and flagnumber
}
\examples{
data(par)
flagSort(flags(par))

}
\seealso{
\code{\link{diffFlags}} calls this function to show the meaning of flags that
are different between two model runs.

\code{\link{read.MFCLFlags}} reads flag settings from a par file.
}
