% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfcl.r
\name{getMFCLversion}
\alias{getMFCLversion}
\title{getMFCLversion}
\usage{
getMFCLversion(file)
}
\arguments{
\item{file}{name of a file to get the MFCL version from.}
}
\value{
the mfcl version
}
\description{
Returns the version of MFCL to be used for this session
}
\details{
The value of \code{file} should be either (1) a path to an MFCL executable
such as \file{mfclo64}, or (2) a path to an MFCL \verb{rep} file such as
\file{plot-10.par.rep}.

The default value of \code{file} looks for \verb{mfclo64} inside the FLR4MFCL
package installation directory.
}
\examples{
repfile <- system.file(package="FLR4MFCL", "extdata/plot-07.par.rep")
getMFCLversion(repfile)

}
