% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfcl.r
\name{gradients}
\alias{gradients}
\title{gradients}
\usage{
gradients(grads = "sorted_gradient.rpt", parnames = "xinit.rpt")
}
\arguments{
\item{grads}{text string of input file name for sorted gradients.}

\item{parnames}{text string of input file name for parameter names.}
}
\value{
data.frame of parameter gradients.
}
\description{
Returns a data.frame of parameters and their gradients in reverse order.
}
\examples{
\dontrun{
gradients()
}

}
