% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{mat2MFCLCor}
\alias{mat2MFCLCor}
\title{Correlation Matrix to Data Frame}
\usage{
mat2MFCLCor(x, names = rownames(x))
}
\arguments{
\item{x}{a correlation matrix (of class \code{matrix}).}

\item{names}{parameter names.}
}
\value{
Data frame containing three columns: \code{Par1}, \code{Par2}, and
\code{Corr}.
}
\description{
Convert a correlation matrix to a three-column data frame.
}
\examples{
data(cormat)
mat2MFCLCor(cormat)
mat2MFCLCor(cormat, names=c("alpha","beta","gamma","delta","epsilon"))

}
\seealso{
\code{\link{read.MFCLCor}} reads a correlation matrix from a file.

\code{\link{corFilter}} filters correlations to show the highest
correlations.

\code{\link{corLabel}} labels correlations and produces a frequency table.
}
