% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.devs.r, R/fnames.r, R/mfcl.r,
%   R/mohns_rho.r, R/plot.r, R/trim.r, R/update.r, R/window.r
\docType{methods}
\name{effort_dev_coffs,MFCLPar,MFCLFrq-method}
\alias{effort_dev_coffs,MFCLPar,MFCLFrq-method}
\alias{mfcl}
\alias{labels}
\alias{fnames}
\alias{check.mfcl.cfg}
\alias{mfcl,MFCLFrq,MFCLPar-method}
\alias{mfcl,MFCLFrq,MFCLIni-method}
\alias{mohns_rho}
\alias{mohns_rho,list-method}
\alias{plot,MFCLLenFreq,ANY-method}
\alias{mfcl-plots}
\alias{trim,MFCLFrqStats-method}
\alias{trim}
\alias{update,MFCLFrq-method}
\alias{window,MFCLFrq-method}
\title{effort_dev_coffs}
\usage{
\S4method{effort_dev_coffs}{MFCLPar,MFCLFrq}(object, frq, ...)

labels(lfile)

fnames(object, ...)

check.mfcl.cfg(dir = ".", spp)

\S4method{mfcl}{MFCLFrq,MFCLPar}(
  frq,
  par,
  outpar = "out.par",
  tag = NULL,
  makepar = F,
  newflags = NULL,
  spp = NULL,
  ...
)

\S4method{mfcl}{MFCLFrq,MFCLIni}(
  frq,
  par,
  outpar = "out.par",
  tag = NULL,
  makepar = F,
  newflags = NULL,
  ...
)

mohns_rho(object, ...)

\S4method{mohns_rho}{list}(object, depletion_method = "SBSBF0", ...)

\S4method{plot}{MFCLLenFreq,ANY}(x, y = "missing", ...)

\S4method{trim}{MFCLFrqStats}(x, ...)

\S4method{update}{MFCLFrq}(object, years, fisheries, multiplier, quantity, ...)

\S4method{window}{MFCLFrq}(
  x,
  start = range(x)["minyear"],
  end = range(x)["maxyear"],
  extend = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class MFCLX.}

\item{frq}{An object of class MFCLFrq.}

\item{\dots}{Additional argument list that might not ever be used.}

\item{lfile}{character string: path to labels.tmp file}

\item{spp}{Character string with 3 letter species code - optional but required to produce the .cfg file if missing.}

\item{par}{An object of class MFCLPar.}

\item{outpar}{The name of the output par file.}

\item{depletion_method}{The method for calculating depletion (default SBSBF0). Obsolte if passing an FLQuants object with 
pre-calculated depletion.}

\item{x}{An object of class MFCLX.}

\item{start}{The start time of the period of interest}

\item{end}{The end time of the period of interest.}

\item{extend}{Logical. If true the start and end times can extend the time series. If false any attempt to extend the series prompts a warning and is ignored.}

\item{names}{An object of either class vector or class data.frame.}

\item{switch}{Optional numeric vector of additional flag settings.}

\item{obj}{An object of class MFCL*.}

\item{...}{Additional argument list that might not ever be used.}
}
\value{
A data.frame of fishery names.

An object of MFCLX.

Creates a text file at the specified location.

A numeric value of mohn's rho.

An updated object of the same class trimmed to the new dimensions.

An updated object of the same class.

An updated object of the same class.
}
\description{
Strip out effort devs from an MFCLPar object including the realisations information from the frq file

read in fishery labels from a labels.tmp file

Add fishery names to an MFCL object

Runs MFCL with defined inputs

Calculate mohns rho for retrospective analyses

Plot MFCL objects.

Trims an object to a new set of smaller dimensions eg. a reduced year range.

Update MFCL objects with new values

Subset MFCL objects between the times start and end.
}
\details{
Trim MFCL objects using named dimensions

The dimensions over which an object can be trimmed will depend on the object itself. 
For example an FLQuant object can be trimmed over the 6 dimensions of an FLQuant (quant/age, year, unit, season, area, iter).
Objects having additional factor levels may allow additional functionality for 'trim'. For example MFCLLenFit() objects 
ma be trimmed on age, length, fishery, year and/or month. Note that where MFCLLenfit objects are trimmed on both age and length
the most restrictive condition is applied.
}
\examples{
\dontrun{
mohns_rho(retros)
}
\dontrun{
plot(MFCLFrq())
}

\dontrun{
trim(MFCLFrq(), year=1990:1995)
}

\dontrun{
update(MFCLFrq())
}

\dontrun{
window(MFCLFrq(), start = 1990, end = 1995)
}

}
\seealso{
\code{\link{MFCLFrq}},  \code{\link{MFCLPar}} and \code{\link{MFCLPar}}

\code{\link{read.MFCLFrq}} and \code{\link{read.MFCLPar}}

\code{\link{SBSBF0}} and \code{\link{MFCLRep}}

\code{\link{MFCLFrq}}, \code{\link{MFCLPar}} and \code{\link{MFCLPar}} and \code{\link{MFCLLenFit}}
}
