% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MFCLEMControl-class}
\alias{MFCLEMControl-class}
\title{An S4 class : Control object for estimation model.}
\description{
An S4 class : Control object for estimation model.
}
\section{Slots}{

\describe{
\item{\code{doitall}}{Description}

\item{\code{tag_fish_rep_rate}}{Description}
}}

