% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{MFCLEMControl}
\alias{MFCLEMControl}
\title{MFCLEMControl}
\usage{
MFCLEMControl(
  doitall = function() {
 },
  tag_fish_rep_rate = 0.9,
  ...
)
}
\description{
Basic constructor for MFCLEMControl class
}
