% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MFCLclasses.r
\docType{class}
\name{MFCLFrq-class}
\alias{MFCLFrq-class}
\title{An S4 class : Representation of a frq input file for MFCL}
\description{
A class comprising an MFCLFrqStats object and an MFCLLenFrq object
}

