% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MFCLFrq2-class}
\alias{MFCLFrq2-class}
\title{An S4 class : Representation of a frq input file for MFCL (based on the MFCLLenFreq2 object)}
\description{
A class comprising an MFCLFrqStats object and an MFCLLenFrq2 object
}
