% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MFCLclasses.r
\docType{class}
\name{MFCLFrqStats-class}
\alias{MFCLFrqStats-class}
\title{An S4 class : Essesntial dimensions and ranges of the frq file.}
\description{
An S4 class : Essesntial dimensions and ranges of the frq file.
}
\section{Slots}{

\describe{
\item{\code{n_regions}}{Number of regions}

\item{\code{n_fisheries}}{Number of fisheries}

\item{\code{n_tag_groups}}{Number of tag groups}

\item{\code{n_recs_yr}}{Number of recruitment events each year}

\item{\code{rec_month}}{Month in which recruitment occurs - 0 means first month of each period}

\item{\code{generic_diffusion}}{Logical}

\item{\code{frq_age_len}}{Logical. Is the frq file age or length based}

\item{\code{frq_version}}{The version of the frq file}

\item{\code{region_size}}{Size of each region relative to first region}

\item{\code{region_fish}}{}

\item{\code{move_matrix}}{}

\item{\code{data_flags}}{}

\item{\code{season_flags}}{}

\item{\code{n_move_yr}}{}

\item{\code{move_weeks}}{}

\item{\code{range}}{}
}}

