% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MFCLLenFit-class}
\alias{MFCLLenFit-class}
\title{An S4 class : Length composition fits from a \file{length.fit} file.}
\description{
An S4 class : Length composition fits from a \file{length.fit} file.
}
\section{Slots}{

\describe{
\item{\code{laa}}{Description}

\item{\code{lenfits}}{Description}

\item{\code{lenagefits}}{Description}

\item{\code{range}}{Description}
}}

