% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MFCLclasses.r
\docType{class}
\name{MFCLLenFreq-class}
\alias{MFCLLenFreq-class}
\title{An S4 class : Length frequency information from the frq file.}
\description{
An S4 class : Length frequency information from the frq file.
}
\section{Slots}{

\describe{
\item{\code{lf_range}}{Range information of the length frequencies}

\item{\code{age_nage}}{I don't know what this is but it's in the frq file}

\item{\code{freq}}{Data frame of length frequency information.}
}}

