% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MFCLLenFreq2-class}
\alias{MFCLLenFreq2-class}
\title{An S4 class : Size frequency information from the frq file.}
\description{
An S4 class : Size frequency information from the frq file.
}
\section{Slots}{

\describe{
\item{\code{lf_range}}{Range information of the length frequencies}

\item{\code{age_nage}}{I don't know what this is but it's in the frq file}

\item{\code{cateffpen}}{Data frame of catch effort and penalty information}

\item{\code{lnfrq}}{Data frame of length frequency information.}

\item{\code{wtfrq}}{Data frame of weight frequency information.}
}}

