% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MFCLLikelihood-class}
\alias{MFCLLikelihood-class}
\title{An S4 class : Likelihood components.}
\description{
An S4 class : Likelihood components.
}
\section{Slots}{

\describe{
\item{\code{bh_steep_contrib}}{Likelihood component}

\item{\code{effort_dev_penalty}}{Likelihood component}

\item{\code{q_dev_pen_fish}}{Likelihood component}

\item{\code{q_dev_pen_fish_grp}}{Likelihood component}

\item{\code{total_length_fish}}{Likelihood component}

\item{\code{length_fish}}{Likelihood component}

\item{\code{total_weight_fish}}{Likelihood component}

\item{\code{weight_fish}}{Likelihood component}

\item{\code{total_catch_fish}}{Likelihood component}

\item{\code{catch_fish}}{Likelihood component}

\item{\code{tag_rel_fish}}{Likelihood component}

\item{\code{survey_index}}{Likelihood component}

\item{\code{age_length}}{Likelihood component}

\item{\code{dimensions}}{Model dimensions}
}}

