% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{MFCLMSEControl}
\alias{MFCLMSEControl}
\title{MFCLMSEControl}
\usage{
MFCLMSEControl(
  hcr = "hcr_threshold",
  hcr_params = c(sbsbf0_min = 0.2, sbsbf0_max = 0.5, out_min = 0.2, out_max = 1),
  itn = 1,
  ess_scalar = 1,
  ...
)
}
\description{
Basic constructor for MFCLMSEControl class
}
