% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{MFCLprojControl}
\alias{MFCLprojControl}
\title{projControl}
\usage{
MFCLprojControl(
  nyears = as.numeric(NULL),
  nsims = as.numeric(NULL),
  avyrs = "",
  fprojyr = as.numeric(NULL),
  controls = data.frame(name = NULL, region = NULL, caeff = NULL, scaler = NULL, ess =
    NULL)
)
}
\description{
Basic constructor for projControl class
}
