% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{MFCLPseudo}
\alias{MFCLPseudo}
\title{MFCLPseudo}
\usage{
MFCLPseudo(
  catcheff = data.frame(year = NULL, month = NULL, fishery = NULL, iter = NULL, data =
    NULL),
  l_frq = data.frame(year = NULL, month = NULL, fishery = NULL, length = NULL, iter =
    NULL, freq = NULL),
  w_frq = data.frame(year = NULL, month = NULL, fishery = NULL, weight = NULL, iter =
    NULL, freq = NULL)
)
}
\description{
Basic constructor for MFCLPseudo class
}
