% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{MFCLPseudoControl}
\alias{MFCLPseudoControl}
\title{MFCLPseudoControl}
\usage{
MFCLPseudoControl(
  catch_sd = 20,
  effort_sd = 20,
  tag_fish_rep_rate = 0.9,
  catch_seed = 16001,
  effort_seed = 17001,
  length_seed = 18001,
  weight_seed = 19001,
  tag_seed = 20001
)
}
\description{
Basic constructor for MFCLPseudoControl class
}
