% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MFCLRegScale-class}
\alias{MFCLRegScale-class}
\title{An S4 class : Size frequency information from the frq file.}
\description{
An S4 class : Size frequency information from the frq file.
}
\section{Slots}{

\describe{
\item{\code{index}}{Description}

\item{\code{dimensions}}{Description}

\item{\code{range}}{Description}
}}

