% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MFCLRep-class}
\alias{MFCLRep-class}
\title{An S4 class : Model information from the rep file.}
\description{
An S4 class : Model information from the rep file.
}
\section{Slots}{

\describe{
\item{\code{fishery_realizations}}{Description}

\item{\code{mean_laa}}{Mean length at age}

\item{\code{mean_waa}}{Mean weight at age}

\item{\code{sd_laa}}{Standard deviation of length at age}

\item{\code{m_at_age}}{Natural mortality}

\item{\code{sel}}{Selectivity}

\item{\code{q_fishery}}{Description}

\item{\code{q_effdev}}{Description}

\item{\code{fm}}{Fishing mortality}

\item{\code{fm_aggregated}}{Fishing mortality aggregated}

\item{\code{popN}}{Population numbers}

\item{\code{rec_region}}{Description}

\item{\code{totalBiomass}}{Total biomass}

\item{\code{totalBiomass_nofish}}{Description}

\item{\code{adultBiomass}}{Description}

\item{\code{adultBiomass_nofish}}{Description}

\item{\code{vulnBiomass}}{Vulnerable biomass}

\item{\code{srr}}{Description}

\item{\code{eq_ssb}}{Equilibrium SSB}

\item{\code{eq_ssb_obs}}{Description}

\item{\code{eq_rec}}{Equilibrium recruitment}

\item{\code{eq_rec_obs}}{Description}

\item{\code{catch_obs}}{Observed catches}

\item{\code{catch_pred}}{Predicted catches}

\item{\code{cpue_obs}}{Observed CPUE}

\item{\code{cpue_pred}}{Predicted CPUE}

\item{\code{eq_biomass}}{Equilibrium biomass}

\item{\code{eq_yield}}{Equilibrium yield}

\item{\code{MSY}}{MSY}

\item{\code{FMSY}}{Fmsy}

\item{\code{BMSY}}{Bmsy}

\item{\code{ABBMSY_ts}}{Description}

\item{\code{FFMSY_ts}}{Description}

\item{\code{ABBMSY}}{Description}

\item{\code{TBBMSY}}{Description}

\item{\code{Fmult}}{Description}

\item{\code{AggregateF}}{Description}
}}

