% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frqTools.r
\name{monthly_to_quarterly}
\alias{monthly_to_quarterly}
\title{monthly_to_quarterly}
\usage{
monthly_to_quarterly(frq, fishery = 15)
}
\arguments{
\item{fishery}{numeric.}
}
\value{
An object of class MFCLFrq
}
\description{
Converts a monthly MFCLFrq object to a quarterly MFCLFrq object
This has been implemented for SP albacore - may not work for other
stocks.
}
\details{
Sums catches over months to quarterly level
Sums length compositions over months to quarterly level

WARNING - this function requires dplyr
}
\examples{
data(frq)
monthly_to_quarterly(frq, fishery=2)

}
