% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagTools.r
\name{parestflag}
\alias{parestflag}
\title{parestflag}
\usage{
parestflag(x, flags, ...)
}
\arguments{
\item{flags}{numeric vector of parest flag number(s).}

\item{par}{object of class MFCLFlag()}
}
\value{
A data frame of sorted flag settings
}
\description{
A simple wrapper function for \code{\link{flagval}}
}
\details{
get parest flag settings
}
\examples{
data(par)
parestflag(par, c(1,20))

}
\seealso{
\code{\link{flagval}} calls this function to retrieve flag settings.

\code{\link{read.MFCLFlags}} reads flag settings from a \verb{par} file.
}
