% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuantMethods.r
\name{qts}
\alias{qts}
\title{qts}
\usage{
qts(quant)
}
\arguments{
\item{quant}{An FLQuant object.}
}
\value{
An object of class FLQuant.
}
\description{
Returns a quarterly time series from a seasonally structured FLQuant.
}
\examples{
data(ple4)
qts(stock.n(ple4))

}
