% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accesorMethods.r
\name{range<-}
\alias{range<-}
\title{range}
\usage{
range(object, ...) <- value

range(object, ...) <- value

range(object, ...) <- value

range(object, ...) <- value

range(object, ...) <- value

range(object, ...) <- value

range(object, ...) <- value
}
\description{
range

range

range

range

range

range

range
}

