% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.MFCLALK.r
\name{read.MFCLALK}
\alias{read.MFCLALK}
\title{read.MFCLALK}
\usage{
read.MFCLALK(alfile, lenfit = NULL)
}
\arguments{
\item{inifile}{A character string giving the name and path of the conditional age length input file to be read.}
}
\value{
An object of class MFCLALK
}
\description{
Reads information from the conditional age length input file and creates an MFCLALK object.
}
\examples{
\dontrun{
read.MFCLALK("C:/R4MFCL/test_data/skj_ref_case/yft.age_length")
}

}
