% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.MFCLPar.r
\name{read.MFCLBiol}
\alias{read.MFCLBiol}
\title{read.MFCLBiol}
\usage{
read.MFCLBiol(parfile, parobj = NULL, first.yr = 1972)
}
\arguments{
\item{parfile:}{A character string giving the name and path of the frq file to be read}

\item{parobj:}{A character string containing the par file. If parobj is NULL the function uses parfile to read in the par file}

\item{first.yr:}{The first year of the input data time series (default values 1972)}
}
\value{
An object of class MFCLBiol
}
\description{
Reads the Biol information from the par file and creates an MFCLBiol object
Unfortunately you still need to supply the first year of the time series.
Other dimensions are interpreted from the data blocks so care should be taken to check
that these have been reproduced correctly
}
\examples{
read.MFCLBiol("C://R4MFCL//test_data//skj_ref_case//11.par")

}
