% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.MFCLCatch.r
\name{read.MFCLCatch}
\alias{read.MFCLCatch}
\title{read.MFCLCatch}
\usage{
read.MFCLCatch(catchfile, dimensions, rnge)
}
\arguments{
\item{catchfile:}{A character string giving the name and path of the catch.rep file to be read}

\item{dimensioins:}{numeric vector of dimensions (as returned from dimensions(par))}

\item{range:}{}
}
\value{
An object of class MFCLCatch
}
\description{
Reads information from the catch.rep file and creates an MFCLCatch object
}
\examples{
\dontrun{
read.MFCLCatch("C:/R4MFCL/test_data/skj_ref_case/catch.rep", dimensions(par))
read.MFCLCatch("/home/robertsc/skj/HCR/run0/catch.rep", dimensions(par))
}

}
