% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.MFCLPseudo.r
\name{read.MFCLCatchSim}
\alias{read.MFCLCatchSim}
\title{read.MFCLCatchSim}
\usage{
read.MFCLCatchSim(
  catch = "catch_sim",
  projfrq = "missing",
  ctrl = "missing",
  historical = TRUE
)
}
\arguments{
\item{catch}{A character string giving the name and path of the catch.sim file to be read.}

\item{projfrq}{The projfrq object used to generate the pseudo data.}

\item{ctrl}{The control object used to generate teh pseudo data.}

\item{historical}{Boolean flag specifying if historical data included.}
}
\value{
An object of class data.frame that can be subset to freq
}
\description{
Reads information from the pseudo generation catch_sim file and creates a data.frame  object.
The object here is to break down the read.MFCLPseudo function in smaller, more manageable code units.
}
