% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{read.MFCLCor}
\alias{read.MFCLCor}
\title{Read Correlation Matrix}
\usage{
read.MFCLCor(x, convert = TRUE, names = NULL)
}
\arguments{
\item{x}{a filename containing a correlation matrix.}

\item{convert}{whether to convert to data frame, using \code{mat2MFCLCor}.}

\item{names}{parameter names, passed to \code{\link{mat2MFCLCor}}.}
}
\value{
Data frame containing three columns: \code{Par1}, \code{Par2}, and
\code{Corr}.
}
\description{
Read a correlation matrix from a file and store as a three-column data frame.
}
\examples{
\dontrun{
read.MFCLCor("skj_pos_hess_cor")
}

}
\seealso{
\code{\link{mat2MFCLCor}} converts a correlation matrix to a data frame.

\code{\link{corFilter}} filters correlations to show the highest
correlations.

\code{\link{corLabel}} labels correlations and produces a frequency table.
}
