% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.MFCLPar.r
\name{read.MFCLFlags}
\alias{read.MFCLFlags}
\title{read.MFCLFlags}
\usage{
read.MFCLFlags(parfile, parobj = NULL, first.yr = NA)
}
\arguments{
\item{parfile}{A character string giving the name and path of the frq file to be read.}

\item{parobj}{A character string containing the par file. If parobj is NULL the function uses parfile to read in the par file.}

\item{first.yr}{The first year of the input data time series (default value NA).}
}
\value{
An object of class MFCLFlags.
}
\description{
Read flag settings from a par file.
}
\note{
The traditional way to read in a par file is using \code{read.MFCLPar} that
imports all parameter values, flags settings, and a variety of other
information.

For the purposes of importing flags, \code{read.MFCLFLags} can be more
practical, being around 20 x faster.

In both cases, the \code{flags()} method can be used to access the
\code{flags} data frame.
}
\examples{
\dontrun{
read.MFCLFlags("C:/R4MFCL/test_data/skj_ref_case/11.par")
}

}
