% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.MFCLfrq.r
\name{read.MFCLFrq}
\alias{read.MFCLFrq}
\title{MFCL frq file reader}
\usage{
read.MFCLFrq(frqfile)
}
\arguments{
\item{frqfile}{A character string giving the name and path of the frq file to be read}
}
\value{
An object of class MFCLFrq
}
\description{
Reads the entire contents of the frq file
}
\examples{
read.MFCLLenFreq(paste(system.file('data', package='FLR4MFCL'), 'skj.frq', sep='//'))rea
}

