% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.MFCLFrq.r
\name{read.MFCLFrqStats}
\alias{read.MFCLFrqStats}
\title{MFCL Frq Statistics}
\usage{
read.MFCLFrqStats(frqfile)
}
\arguments{
\item{frqfile}{A character string giving the name and path of the frq file to be read}
}
\value{
An object of class MFCLFrqStats
}
\description{
Extracts the essential ranges and dimensions from the frq file
}
\examples{
read.MFCLFrqStats("C://R4MFCL//test_data//skj_ref_case//skj.frq")

}
