% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.MFCLfrq.r
\name{read.MFCLLenFreq}
\alias{read.MFCLLenFreq}
\title{MFCL Length Frequency}
\usage{
read.MFCLLenFreq(frqfile)
}
\arguments{
\item{frqfile}{A character string giving the name and path of the frq file to be read}
}
\value{
An object of class MFCLLenFreq
}
\description{
Extracts the length frequency data from the frq file
}
\examples{
read.MFCLLenFreq("C://R4MFCL//test_data//skj_ref_case//skj.frq")
}

