% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.MFCLFrq.r
\name{read.MFCLLenFreq2}
\alias{read.MFCLLenFreq2}
\title{MFCL Length Frequency2 - MUFDAGR friendly}
\usage{
read.MFCLLenFreq2(frqfile)
}
\arguments{
\item{frqfile}{A character string giving the name and path of the frq file to be read}
}
\value{
An object of class MFCLLenFreq
}
\description{
Extracts the length frequency data from Mufdagr output
}
\examples{
read.MFCLLenFreq("C://R4MFCL//test_data//skj_ref_case//skj.frq")

}
