% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.MFCLLikelihood.r
\name{read.MFCLLikelihood}
\alias{read.MFCLLikelihood}
\title{Read MFCLLikelihood}
\usage{
read.MFCLLikelihood(tpofile = "test_plot_output")
}
\arguments{
\item{tpofile}{file containing the likelihood component values.}
}
\value{
An object of class \code{MFCLLikelihood}.
}
\description{
Read likelihood components from the \file{test_plot_output file} and create
an \code{MFCLLikelihood} object.
}
\examples{
\dontrun{
read.MFCLLikelihood()
read.MFCLRep("/home/roberts/skj/HCR/run0/test_plot_output")
}

}
