% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.MFCLPseudo.r
\name{read.MFCLPseudo}
\alias{read.MFCLPseudo}
\title{Read MFCLPseudo}
\usage{
read.MFCLPseudo(
  catch = "missing",
  effort = "missing",
  lw_sim = "missing",
  projfrq = "missing",
  ctrl = "missing",
  historical = TRUE,
  ii = 1
)
}
\arguments{
\item{catch}{A character string giving the name and path of the catch.sim file to be read.}

\item{effort}{A character string giving the name and path of the effort.sim file to be read.}

\item{lw_sim}{A character string giving the name and path of the test_lw_sim (or test_lw_sim_alt) file to be read.}

\item{projfrq}{The projection *.frq object from which the pseudo data were generated.}

\item{ctrl}{The corresponding projection control object.}
}
\value{
An object of class MFCLPseudo
}
\description{
Reads information from the pseudo generation files and creates an MFCLPseudo object.
}
