% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.MFCLPseudo.r
\name{read.MFCLPseudoCatchEffort}
\alias{read.MFCLPseudoCatchEffort}
\title{'read.MFCLPseudoCatchEffort()' function for FLR4MFCL}
\usage{
read.MFCLPseudoCatchEffort(
  catch = "catch_sim",
  effort = "effort_sim",
  projfrq = projfrq,
  ctrl = "missing",
  historical = FALSE
)
}
\arguments{
\item{catch}{A character string for the 'catch_sim' input file name.}

\item{effort}{A character string for the 'effort_sim' input file name.}

\item{projfrq}{An \linkS4class{MFCLFrq} object for the projection period over which the pseudo data have been generated.}

\item{ctrl}{An \linkS4class{MFCLMSEControl}}

\item{historical}{Boolean TRUE or FALSE If the simulated data include the historical period or just the projecion period.}
}
\value{
An object of class \linkS4class{MFCLPseudo}.
}
\description{
Reads in simulated catch and effort data from the MFCL generated files 'catch_sim' and 'effort_sim' and
returns a partially complete object of MFCLPseudo. The size composition slots of the MFCLPseudo object are not filled.
}
\details{
There are now several functions for reading pseudo data :
\itemize{
  \item read.MFCLPseudo - reads both catch, effort and size comps. Works when you only have either pseudo length or weight composition data (e.g. skipjack). 
  \item read.MFCLCatchEffort - only reads in the simulated catch and effort data
  \item read.MFCLPseudoSizeComp - only reads the size comp data from test_lw_sim but works when you have both length and weight data.
}
}
\examples{
\dontrun{
# Expanding an MFCLFrq, e.g. that was used from in an assessment
pseudo <- read.MFCLPseudoCatchEffort('catch_sim', 'effort_sim', projfrq, mseCtrl, historical=FALSE)
}
}
\seealso{
\code{\link{MFCLprojContrl}} \code{\link{MFCLFrq}} \code{\link{MFCLPar}}
}
