% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.MFCLPar.r
\name{read.MFCLRec}
\alias{read.MFCLRec}
\title{read.MFCLRec}
\usage{
read.MFCLRec(parfile, parobj = NULL, first.yr = 1972)
}
\arguments{
\item{parfile:}{A character string giving the name and path of the frq file to be read}

\item{parobj:}{A character string containing the par file. If parobj is NULL the function uses parfile to read in the par file}

\item{first.yr:}{The first year of the input data time series (default value 1972)}
}
\value{
An object of class MFCLRec
}
\description{
Reads the recruitment information from the par file and creates an MFCLec object
}
\examples{
read.MFCLRec("C://R4MFCL//test_data//skj_ref_case//11.par")

}
