% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.MFCLRep.r
\name{read.MFCLRep}
\alias{read.MFCLRep}
\title{read.MFCLRep}
\usage{
read.MFCLRep(repfile)
}
\arguments{
\item{inifile:}{A character string giving the name and path of the .rep file to be read}
}
\value{
An object of class MFCLRep
}
\description{
Reads information from the .rep file and creates an MFCLRep object
}
\examples{
read.MFCLIni("C:/R4MFCL/test_data/skj_ref_case/plot-out.par.rep")
read.MFCLIni("/home/roberts/skj/HCR/run0/plot-out.par.rep")

}
