% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.MFCLRepTag.r
\name{read.MFCLRepTag}
\alias{read.MFCLRepTag}
\title{read.MFCLRepTag}
\usage{
read.MFCLRepTag(tagrep)
}
\arguments{
\item{tagrep}{A character string giving the name and path of the tag.rep file to be read.}
}
\value{
An object of class character data.frame.
}
\description{
Reads information from the tag file and creates object.
}
\examples{
\dontrun{
read.MFCLRepTag("/home/robertsc/MSE/YFT/OM_development/OM_grid_fits/2026_OMs_test/OM0_growth_ini_repair/tag.rep")
}

}
