% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.simulated.r
\name{read.MFCLSimulatedNatAge}
\alias{read.MFCLSimulatedNatAge}
\title{read.simulated}
\usage{
read.MFCLSimulatedNatAge(
  file = "simulated_numbers_at_age",
  data = "N",
  dimensions = NULL,
  rnge = NULL,
  nssns = 4
)
}
\arguments{
\item{file}{A character string giving the name and path of the simulated_numbers_at_age file to be read.}

\item{data}{A character (either "N" or "R") to specify whether termninal N values or recruitment values are to be read in.}

\item{dimensions}{numeric vector of dimensions (as returned from dimensions(par)).}

\item{range}{}
}
\value{
An object of class FLQuant
}
\description{
Reads information from the simulated_numbers_at_age file (terminal N or recruitment) and creates an FLQuant object
}
