% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.MFCLTag.r
\name{read.MFCLTag}
\alias{read.MFCLTag}
\title{read.MFCLTag}
\usage{
read.MFCLTag(tagfile)
}
\arguments{
\item{tagfile:}{A character string giving the name and path of the tag file to be read}
}
\value{
An object of class character vector
}
\description{
Reads information from the tag file and creates  object
}
\examples{
read.MFCLTag("C:/R4MFCL/test_data/skj_ref_case/skj.tag")

}
