% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.MFCLPar.r
\name{read.MFCLTagRep}
\alias{read.MFCLTagRep}
\title{read.MFCLTagRep}
\usage{
read.MFCLTagRep(parfile, parobj = NULL, first.yr = 1972)
}
\arguments{
\item{parfile:}{A character string giving the name and path of the frq file to be read}

\item{parobj:}{A character string containing the par file. If parobj is NULL the function uses parfile to read in the par file}
}
\value{
An object of class MFCLTagRep
}
\description{
Reads the TAaRep information from the par file and creates an MFCLTagRep object
}
\examples{
read.MFCLTagRep("C:/R4MFCL/test_data/skj_ref_case/11.par")

}
