% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.MFCLVar.R
\name{read.MFCLVar}
\alias{read.MFCLVar}
\title{Read Var File}
\usage{
read.MFCLVar(varfile)
}
\arguments{
\item{varfile}{filename ending with \verb{.var}.}
}
\value{
Vector of reference point estimates.
}
\description{
Read reference point estimates from the \verb{.var} file.
}
\note{
The \verb{.var} file is produced by MFCL as a result of delta-method
analysis, following Hessian calculations.
}
\section{Warning}{

The calculation of the Frecent/Fmsy in the \verb{.var} file is based on a
harvest rate approach, comparing catch in tonnes with biomass in tonnes.
Importantly, this is different from the calculation of Frecent/Fmsy in the
\verb{.rep} file, which uses an instantaneous mortality rate, comparing catch
in numbers with population size in numbers.

The choice between reading Frecent/Fmsy from the \verb{.rep} file or the
\verb{.var} file should therefore be based on an informed decision,
understanding the differences between the two approaches.
}

\examples{
\dontrun{
read.MFCLVar("yft.var")
}

}
\seealso{
\code{\link{read.MFCLRep}}.
}
