% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.MFCLfrq.r
\name{read.MFCLVitalStats}
\alias{read.MFCLVitalStats}
\title{MFCL Vital Statistics}
\usage{
read.MFCLVitalStats(frqfile)
}
\arguments{
\item{frqfile}{A character string giving the name and path of the frq file to be read}

\item{cex}{character expansion for the text (\code{"hsv"} or \code{"cluster"})}

\item{mar}{margin paramaters; vector of length 4 (see \code{\link[graphics]{par}})}
}
\value{
An object of class MFCLVitalStats
}
\description{
Extracts the essential ranges and dimensions from the frq file
}
\examples{
read.MFCLVitalStats("C://R4MFCL//test_data//skj_ref_case//skj.frq")
}

