% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.MFCLWgtFit.r
\name{read.MFCLWgtFit}
\alias{read.MFCLWgtFit}
\title{read.MFCLWgtFit}
\usage{
read.MFCLWgtFit(wffile, get_wgtage = FALSE)
}
\arguments{
\item{inifile}{A character string giving the name and path of the weight.fit file to be read.}
}
\value{
An object of class MFCLWgtFit
}
\description{
Reads information from the weight.fit file and creates an MFCLWgtFit object.
}
\examples{
\dontrun{
read.MFCLWgtFit("C:/R4MFCL/test_data/skj_ref_case/weight.fit")
}

}
