% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagTools.r
\name{regionflag}
\alias{regionflag}
\title{regionflag}
\usage{
regionflag(x, flags, ...)
}
\arguments{
\item{flags}{numeric vector of region flag number(s).}

\item{par}{object of class MFCLFlag()}
}
\value{
A data frame of sorted flag settings
}
\description{
A simple wrapper function for \code{\link{flagval}}
}
\details{
get region flag settings
}
\examples{
data(par)
regionflag(par, c(1))

}
\seealso{
\code{\link{flagval}} calls this function to retrieve flag settings.

\code{\link{read.MFCLFlags}} reads flag settings from a \verb{par} file.
}
