% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repMethods.r
\docType{methods}
\name{SBF0Alt}
\alias{SBF0Alt}
\alias{SBF0Alt,MFCLRep,numeric,missing-method}
\alias{SBF0Alt,MFCLRep,numeric,numeric-method}
\title{SBF0Alt}
\usage{
SBF0Alt(rep, years, lag, ...)

\S4method{SBF0Alt}{MFCLRep,numeric,missing}(rep, years, rolling = TRUE,
  combine_areas = TRUE)

\S4method{SBF0Alt}{MFCLRep,numeric,numeric}(rep, years, lag,
  rolling = TRUE, combine_areas = TRUE)
}
\arguments{
\item{rep}{An object of class MFCLRep.}

\item{years}{Either a vector of years over which to calculate the mean over, or a single numeric vector giving the number of years over which to calculate the mean.}

\item{lag}{If \code{years} is a single numeric vector, the number of years between the final year in the rep file and final year over which to calculate the mean.}

\item{rolling}{TRUE (a rolling mean, i.e. a moving average) or FALSE (the years over which to calculate the mean are fixed).  Default is TRUE.}

\item{combine_areas}{A logical argument. Should the areas (regions) be combined. Default is TRUE.}
}
\value{
An FLQuant object with annual mean SBF0 values.
}
\description{
Alternative method for calculating mean SBF0 for use in the SB/SBF0 calculation.
}
\details{
The method calculates the mean SBF0 over a year range. The year range can be specified in two ways, by either specifying a year range with the \code{years} argument, or via the number of years over which to calculate the mean (the \code{years} argument as a single value) and the number of years between the final year and the last year of the year range (\code{lag}).
If rolling is FALSE, the mean is calculated for the year range only. If rolling is TRUE then the year range is relative to the final year and shifts about accordingly.

For example, if the final year is 2042, the year range is 2011:2015 and rolling is FALSE, the returned FLQuant has the mean SBF0 over the years 2011:2015 for all years.
If rolling is TRUE, then year range 2011:2015 is used to calculate the SBF0 for 2042, the year range 2010:2014 is used for 2041 and so on.
If years is a single value of 5, and lag is 10, then the mean SBF0 in 2042 is calculated over the year range 2028:2032 (i.e. over years). If rolling is TRUE, the mean in SBF) in 2041 is calculated over the range 2027:2031 and so on.
}
\examples{
\dontrun{
# Mean over a constant year range of 2011:2015
SBF0Alt(rep, years=2011:2015, rolling=FALSE)
# Mean over rolling year range, with the final year being the mean of 2011:2015, the second to last year being mean od 2010:2014 etc.
SBF0Alt(rep, years=2011:2015, rolling=TRUE)
# Mean over a constant year range of defined as number of years in the range and number or years from the final year.
# e.g. if the final year was 2042, return the mean over 2028:2032
SBF0Alt(rep, years=5, lag=10, rolling=FALSE)
# Mean over a rolling year range, e.g. value in 2042 is the mean over 2028:2032, value in 2041 is mean over 2027:2031 etc.
SBF0Alt(rep, years=5, lag=10, rolling=FALSE)
}

}
\seealso{
\code{\link{read.MFCLRep}}
}
