% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.r
\docType{methods}
\name{subset}
\alias{subset}
\alias{subset,MFCLFrq-method}
\title{subset}
\usage{
subset(x, ...)

\S4method{subset}{MFCLFrq}(x, ...)
}
\arguments{
\item{x}{An object of class MFCLFrq.}

\item{...}{Additional argument list.}
}
\value{
A data.frame of the selected information.
}
\description{
Subsets a freq(frq) for specific groupings.
}
