% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.r
\name{summary,MFCLLikelihood-method}
\alias{summary,MFCLLikelihood-method}
\title{Summary MFCLLikelihood}
\usage{
\S4method{summary}{MFCLLikelihood}(object)
}
\arguments{
\item{object}{An object of class MFCLLikelihood.}
}
\value{
Data frame containing aggregated likelihood components.
}
\description{
Produce a summary of likelihood components.
}
\examples{
\dontrun{
ll <- read.MFCLLikelihood()
summary(ll)
}
}
\seealso{
\code{\link{read.MFCLLikelihood}}, \code{\link{MFCLLikelihood-class}}.
}
