% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.r, R/write.ALK.r,
%   R/write.MFCLFrq.r, R/write.MFCLIni.r, R/write.MFCLPar.r,
%   R/write.MFCLRegScale.r, R/write.MFCLTag.r, R/write.MFCLTagProj.r
\docType{methods}
\name{write}
\alias{write}
\alias{write,MFCLALK-method}
\alias{write,MFCLFrqStats-method}
\alias{write,MFCLLenFreq-method}
\alias{write,MFCLFrq-method}
\alias{write,MFCLIni-method}
\alias{write,MFCLPar-method}
\alias{write,MFCLRegScale-method}
\alias{write,MFCLTag-method}
\alias{write,MFCLTagProj-method}
\title{write}
\usage{
write(x, file, append = F, ...)

\S4method{write}{MFCLALK}(x, file, append = F, ...)

\S4method{write}{MFCLFrqStats}(x, file, append = F, ...)

\S4method{write}{MFCLLenFreq}(x, file, append = T, ...)

\S4method{write}{MFCLFrq}(x, file, append = F, ...)

\S4method{write}{MFCLIni}(x, file, append = F, ...)

\S4method{write}{MFCLPar}(x, file, append = F, ...)

\S4method{write}{MFCLRegScale}(x, file, append = F, ...)

\S4method{write}{MFCLTag}(x, file, append = F, ...)

\S4method{write}{MFCLTagProj}(x, file, append = F, ...)
}
\arguments{
\item{x}{An object of class MFCL eg. MFCLFrq, MFCLPar, etc.}

\item{file}{The name and path of the file to be written}

\item{append}{If True append to existing file, If False overwrite any existing file}

\item{...}{Additional argument list that might not ever
be used.}
}
\value{
Creates a text file at the specified location.
}
\description{
Writes MFCL objects to a text file
}
\examples{
\dontrun{
write(MFCLFrqStats(), "file.txt")
}
}
\seealso{
\code{\link{read.MFCLFrq}} and \code{\link{read.MFCLPar}}
}
