% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdultJuvenileF.R
\name{AdultJuvenileF}
\alias{AdultJuvenileF}
\title{Adult and Juvenile F}
\usage{
AdultJuvenileF(rep, par)
}
\arguments{
\item{rep}{object of class MFCLRep.}

\item{par}{object of class MFCLPar.}
}
\value{
Data frame containing four columns:
\item{year}{year}
\item{area}{region, including \code{"all"}}
\item{stage}{stage, \code{"adult"} and \code{"juvenile"}}
\item{f}{fishing mortality rate, annual}
}
\description{
Calculate adult and juvenile fishing mortality.
}
\note{
Adult fishing mortality is calculated as the weighted average fishing
mortality, weighted by the maturity ogive.

Juvenile fishing mortality is calculated as the weighted average fishing
mortality, weighted by the inverse \code{(1-p)} of the maturity ogive.
}
\seealso{
\code{\link{fm}} and \code{\link{fm_aggregated}} are used to access the
fishing mortalities.

\code{\link{mat}} is used to access the maturity ogive.
}
