% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagTools.r
\name{flagDiff}
\alias{flagDiff}
\title{Flag Diff}
\usage{
flagDiff(par1, par2)
}
\arguments{
\item{par1}{An object of class MFCLPar.}

\item{par2}{An object of class MFCLPar.}
}
\value{
A data frame of flag settings for par1 and par2.
}
\description{
Show flag differences between two par objects/files.
}
\examples{
data(par)
par1 <- par2 <- par
flags(par2)[20,"value"] <- 12
flagDiff(par1, par2)

}
