% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagTools.r
\name{flagDiff}
\alias{flagDiff}
\title{flagDiff}
\usage{
flagDiff(par1, par2)
}
\arguments{
\item{par}{A object of class MFCLPar}
}
\value{
A data frame of flag settings for par1 and par2
}
\description{
flag differences between two par files
}
\examples{
flagDiff(par, par)

}
