% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagTools.r
\name{flagMeaning}
\alias{flagMeaning}
\title{Flag Meaning}
\usage{
flagMeaning(flags, flaglist = NULL)
}
\arguments{
\item{flags}{is a data frame or object of class \code{MFCLPar} or
\code{MFCLFlags}.}

\item{flaglist}{optional filename to use instead of the built-in
\file{flaglist.csv} lookup table.}
}
\value{
A data frame with the same columns as \code{flags} plus a column called
\code{meaning}.
}
\description{
Show the meaning of flags, based on a lookup table.
}
\examples{
data(par)
par1 <- par2 <- par

# Different flag value
flags(par2)[20,"value"] <- 12
flagDiff(par1, par2)

# Example where flag is specified in par1 but not in par2
flags(par1) <- rbind(flags(par1), c(-10269, 1, 1))
flagDiff(par1, par2)             # default is to show par2 as NA
flagDiff(par1, par2, all=FALSE)  # all=FALSE omits such comparisons

}
\seealso{
\code{\link{read.MFCLFlags}}, \code{\link{flagDiff}}.
}
