% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagTools.r
\name{flagSummary}
\alias{flagSummary}
\title{flag summary}
\usage{
flagSummary(par, type)
}
\arguments{
\item{par}{An object of class MFCLPar}

\item{type}{A character string specifying the MFCL User Guide section}
}
\value{
A data frame of flag settings
}
\description{
flag settings summarised by MFCL User Guide sections
}
\examples{
data(par)
flagSummary(par, 'projection')

}
