% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.r
\docType{methods}
\name{realisations}
\alias{realisations}
\title{realisations}
\usage{
realisations(object, ...)
}
\arguments{
\item{object}{An object of class MFCLLenFreq.}

\item{\dots}{Currently ignored.}
}
\value{
An object of class data.frame.
}
\description{
Returns the unique fishery realisations. 
Essentially this is the contents of the \code{freq} slot of an \code{\link{MFCLFrq}} object 
without the replications for each length or weight category.
}
\examples{
\dontrun{
realisations(frq)
}
}
\seealso{
\code{\link{read.MFCLFrq}}
}
