% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfcl.r
\name{getMFCLversion}
\alias{getMFCLversion}
\title{getMFCLversion}
\usage{
getMFCLversion()
}
\value{
the mfcl version
}
\description{
Returns the version of MFCL to be used for this session
}
\examples{
getMFCLversion()

}
