% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcrs.R
\name{hcr_threshold}
\alias{hcr_threshold}
\alias{hcr_constant}
\alias{hcr_threshold_constrained}
\title{hcr_threshold}
\usage{
hcr_threshold(sbsbf0, params)

hcr_constant(sbsbf0, params)

hcr_threshold_constrained(sbsbf0, params, reference_out)
}
\arguments{
\item{sbsbf0}{Numeric vector of SBSBF0.}

\item{params}{Numeric vector of HCR parameters.}

\item{reference_out}{The reference level that the constraint is applied to.}
}
\value{
Numberic vector of the HCR output
}
\description{
Typical HCR function with the parameters sbsbf0_min, sbsbf0_max, out_min, out_max passed as a vector.
The default HCR for the MFCLMSEControl class.

An HCR function that just returns the same output irrespective of the SBSBF0 input

Like the typical HCR function with the parameters but with a constraint on how much the output is allowed to change from a reference output.
Parameters are 'sbsbf0_min', 'sbsbf0_max', 'out_min', 'out_max', 'max_change_up', 'max_change_down'.
There is an additional argument 'reference_out' that is the reference output, i.e. the value to which to the new output is compared.
}
