% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.devs.r, R/mfcl.r, R/mohns_rho.r,
%   R/plot.r, R/trim.r, R/update.r, R/window.r
\docType{methods}
\name{effort_dev_coffs,MFCLPar,MFCLFrq-method}
\alias{effort_dev_coffs,MFCLPar,MFCLFrq-method}
\alias{mfcl}
\alias{check.mfcl.cfg}
\alias{mfcl,MFCLFrq,MFCLPar-method}
\alias{mfcl,MFCLFrq,MFCLIni-method}
\alias{mohns_rho}
\alias{mohns_rho,list-method}
\alias{plot,MFCLLenFreq,ANY-method}
\alias{trim,MFCLFrqStats-method}
\alias{trim,MFCLRep-method}
\alias{update,MFCLFrq-method}
\alias{window,MFCLFrq-method}
\title{effort_dev_coffs}
\usage{
\S4method{effort_dev_coffs}{MFCLPar,MFCLFrq}(object, frq, ...)

check.mfcl.cfg(dir = ".", spp)

\S4method{mfcl}{MFCLFrq,MFCLPar}(
  frq,
  par,
  outpar = "out.par",
  tag = NULL,
  makepar = F,
  newflags = NULL,
  spp = NULL,
  ...
)

\S4method{mfcl}{MFCLFrq,MFCLIni}(
  frq,
  par,
  outpar = "out.par",
  tag = NULL,
  makepar = F,
  newflags = NULL,
  ...
)

mohns_rho(object, ...)

\S4method{mohns_rho}{list}(object, depletion_method = "SBSBF0", ...)

\S4method{plot}{MFCLLenFreq,ANY}(x, y = "missing", ...)

\S4method{trim}{MFCLFrqStats}(x, ...)

\S4method{trim}{MFCLRep}(x, ...)

\S4method{update}{MFCLFrq}(object, years, fisheries, multiplier, quantity, ...)

\S4method{window}{MFCLFrq}(
  x,
  start = range(x)["minyear"],
  end = range(x)["maxyear"],
  extend = FALSE,
  ...
)
}
\arguments{
\item{frq}{An object of class MFCLFrq.}

\item{...}{Additional argument list that might not ever
be used.}

\item{spp}{Character string with 3 letter species code - optional but required to produce the .cfg file if missing.}

\item{par}{An object of class MFCLPar.}

\item{outpar}{The name of the output par file.}

\item{depletion_method}{The method for calculating depletion (default SBSBF0).}

\item{object:}{An object of class MFCLPar}

\item{frq:}{The corresponding MFCLFrq for the MFCLPar}

\item{switch}{Optional numeric vector of additional flag settings.}

\item{obj}{An object of class list containing the rep files from the final assessment and each retrospective peel.}

\item{obj:}{An object of class MFCLX.}

\item{x:}{An object of class MFCLX.}

\item{start:}{The start time of the period of interest}

\item{end:}{The end time of the period of interest.}

\item{extend:}{logical. If true the start and end times can extend the time series. If false any attempt to extend the series prompts a warning and is ignored.}
}
\value{
Creates a text file at the specified location.

A numeric value of mohn's rho.

An updated object of the same class

An updated object of the same class

An updated object of the same class
}
\description{
Strip out effort devs from an MFCLPar object including the realisations information from the frq file

Runs MFCL with defined inputs

Calculate mohns rho for retrospective analyses

Plot MFCL objects

trim MFCL objects using named dimensions

Update MFCL objects with new values

subset MFCL objects between the times start and end.
}
\examples{
\dontrun{
mohns_rho(retros)
}
\dontrun{
plot(MFCLFrq())
}

trim(MFCLFrq(), year=1990:1995)

update(MFCLFrq())

window(MFCLFrq(), start = 1990, end = 1995)

}
\seealso{
\code{\link{MFCLFrq}},  \code{\link{MFCLPar}} and \code{\link{MFCLPar}}

\code{\link{read.MFCLFrq}} and \code{\link{read.MFCLPar}}

\code{\link{SBSBF0}} and \code{\link{MFCLRep}}

\code{\link{MFCLFrq}},  \code{\link{MFCLPar}} and \code{\link{MFCLPar}}

\code{\link{read.MFCLFrq}} and \code{\link{read.MFCLPar}}

\code{\link{MFCLFrq}},  \code{\link{MFCLPar}} and \code{\link{MFCLPar}}
}
