% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MFCLclasses.r
\docType{class}
\name{MFCLFrq-class}
\alias{MFCLFrq-class}
\title{An S4 class : Representation of a frq input file for MFCL}
\description{
An S4 class : Representation of a frq input file for MFCL
}
\section{Slots}{

\describe{
\item{\code{lf_range}}{Range information of the length frequencies}
}}

