% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MFCLFrqStats-class}
\alias{MFCLFrqStats-class}
\title{An S4 class : Essential dimensions and ranges of the frq file.}
\description{
An S4 class : Essential dimensions and ranges of the frq file.
}
\section{Slots}{

\describe{
\item{\code{n_regions}}{Number of regions}

\item{\code{n_fisheries}}{Number of fisheries}

\item{\code{n_tag_groups}}{Number of tag groups}

\item{\code{n_recs_yr}}{Number of recruitment events each year}

\item{\code{rec_month}}{Month in which recruitment occurs - 0 means first month of each period}

\item{\code{generic_diffusion}}{Logical}

\item{\code{frq_age_len}}{Logical. Is the frq file age or length based}

\item{\code{frq_version}}{The version of the frq file}

\item{\code{region_size}}{Size of each region relative to first region}

\item{\code{region_fish}}{Description}

\item{\code{move_matrix}}{Description}

\item{\code{data_flags}}{Description}

\item{\code{season_flags}}{Description}

\item{\code{n_move_yr}}{Description}

\item{\code{move_weeks}}{Description}

\item{\code{range}}{Description}
}}

