% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MFCLIni-class}
\alias{MFCLIni-class}
\title{An S4 class : Model information from the ini file.}
\description{
An S4 class : Model information from the ini file.
}
\section{Slots}{

\describe{
\item{\code{dimensions}}{Description}

\item{\code{range}}{Description}

\item{\code{tag_shed_rate}}{Description}

\item{\code{tag_fish_rep_rate}}{Description}

\item{\code{tag_fish_rep_grp}}{Description}

\item{\code{tag_fish_rep_flags}}{Description}

\item{\code{tag_fish_rep_target}}{Description}

\item{\code{tag_fish_rep_pen}}{Description}

\item{\code{rep_rate_dev_coffs}}{Description}

\item{\code{m}}{Description}

\item{\code{m_devs_age}}{Description}

\item{\code{log_m}}{Description}

\item{\code{mat}}{Description}

\item{\code{mat_at_length}}{Description}

\item{\code{growth}}{Description}

\item{\code{richards}}{Description}

\item{\code{growth_var_pars}}{Description}

\item{\code{n_mean_constraints}}{Description}

\item{\code{growth_devs_age}}{Description}

\item{\code{growth_curve_devs}}{Description}

\item{\code{growth_devs_cohort}}{Description}

\item{\code{season_growth_pars}}{Description}

\item{\code{len_bias_pars}}{Description}

\item{\code{common_len_bias_pars}}{Description}

\item{\code{common_len_bias_coffs}}{Description}

\item{\code{control_flags}}{Description}

\item{\code{move_map}}{Description}

\item{\code{diff_coffs}}{Description}

\item{\code{xdiff_coffs}}{Description}

\item{\code{y1diff_coffs}}{Description}

\item{\code{y2diff_coffs}}{Description}

\item{\code{zdiff_coffs}}{Description}

\item{\code{diff_coffs_mat}}{Description}

\item{\code{diff_coffs_age_ssn}}{Description}

\item{\code{diff_coffs_age_period}}{Description}

\item{\code{diff_coffs_age}}{Description}

\item{\code{diff_coffs_nl}}{Description}

\item{\code{diff_coffs_priors}}{Description}

\item{\code{diff_coffs_age_priors}}{Description}

\item{\code{diff_coffs_nl_priors}}{Description}

\item{\code{region_rec_var}}{Description}

\item{\code{region_pars}}{Description}

\item{\code{ini_version}}{Description}

\item{\code{region_flags}}{Description}

\item{\code{age_pars}}{Description}

\item{\code{rec_dist}}{Description}

\item{\code{lw_params}}{Description}

\item{\code{sv}}{Description}

\item{\code{sd_length_at_age}}{Description}

\item{\code{sd_length_dep}}{Description}
}}

