% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.r, R/parMethods.r
\docType{methods}
\name{as.MFCLIni}
\alias{as.MFCLIni}
\alias{n_fisheries,MFCLPar-method}
\alias{laa,MFCLBiol-method}
\alias{laa}
\alias{aal,MFCLBiol-method}
\alias{aal}
\alias{waa}
\alias{waa,MFCLBiol-method}
\alias{SPR0}
\alias{SPR0,MFCLBiol-method}
\alias{YPR}
\alias{YPR,MFCLRep,MFCLPar-method}
\alias{recYears}
\alias{recYears,MFCLPar-method}
\title{as.MFCLIni}
\usage{
as.MFCLIni(object, ...)

\S4method{n_fisheries}{MFCLPar}(object, ...)

\S4method{laa}{MFCLBiol}(object, ages = NULL, ...)

\S4method{aal}{MFCLBiol}(object, lengths = seq(0, 108, by = 2))

waa(object, ...)

\S4method{waa}{MFCLBiol}(object, ...)

SPR0(par, ...)

\S4method{SPR0}{MFCLBiol}(par, ...)

YPR(object, par, ...)

\S4method{YPR}{MFCLRep,MFCLPar}(object, par, fminyr = NULL, fmaxyr = NULL, scalars = NULL, ...)

recYears(par, ...)

\S4method{recYears}{MFCLPar}(par, nrecsyr = 4, ...)
}
\arguments{
\item{rep:}{An object of class MFCLRep}

\item{object:}{An object of class MFCLRep}

\item{par:}{An object of class MFCLPar}
}
\value{
An MFCLIni object.

A vector of lengths at age.

A vector of weights at age.

An FLQuant.

A dataframe.

A vector.
}
\description{
Convert an MFCLPar object to an MFCLIni object

Calculates length at age from Von Bertalanffy parameters in the par file

Calculates weight at age from Von Bertalanffy parameters and length weight params in the par file

Calculates spawners per recruit at zero fishing

Calculates yield per recruit and spawners per recruit - in kilos

Returns the year range over which the SRR is calculated given the flag settings from af199 and af200
}
