% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.MFCLLenFit.r
\name{read.MFCLLenFit}
\alias{read.MFCLLenFit}
\title{read.MFCLLenFit}
\usage{
read.MFCLLenFit(lffile)
}
\arguments{
\item{inifile:}{A character string giving the name and path of the length.fit file to be read}
}
\value{
An object of class MFCLLenFit
}
\description{
Reads information from the length.fit file and creates an MFCLLenFit object
}
\examples{
read.MFCLLenFit("C://R4MFCL//test_data//skj_ref_case//length.fit")

}
