% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.MFCLPseudo.r
\name{read.MFCLPseudo}
\alias{read.MFCLPseudo}
\title{read.MFCLPseudo}
\usage{
read.MFCLPseudo(
  catch = "missing",
  effort = "missing",
  lw_sim = "missing",
  projfrq = "missing",
  ctrl = "missing",
  historical = TRUE
)
}
\arguments{
\item{catch:}{A character string giving the name and path of the catch.sim file to be read}

\item{effort:}{A character string giving the name and path of the effort.sim file to be read}

\item{lw_sim:}{A character string giving the name and path of the lw_sim file to be read}

\item{range:}{}
}
\value{
An object of class MFCLPseudo
}
\description{
Reads information from the pseudo generation files and creates an MFCLPseudo object.
This method is a complicated mess and really needs to be simplified.
}
