% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.MFCLRegScale.r
\name{read.MFCLRegScale}
\alias{read.MFCLRegScale}
\title{read.MFCLRegScale}
\usage{
read.MFCLRegScale(file, seasons = 4, ...)
}
\arguments{
\item{file}{A character string giving the name and path of the regional scaling input file to be read.}

\item{seasons}{The number of seasons - defaults to 4. This should be 1 for an annual model.}
}
\value{
An object of class MFCLRegScale
}
\description{
Reads information from the regional scaling input file and creates an MFCLRegScale object.
}
\examples{
\dontrun{
read.MFCLRegScale("C:/R4MFCL/test_data/skj_ref_case/yft.reg_scaling")
}

}
