% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagTools.r
\name{recPeriod}
\alias{recPeriod}
\title{recruitment periods}
\usage{
recPeriod(
  par,
  af199 = NULL,
  af200 = NULL,
  pf232 = NULL,
  pf233 = NULL,
  show = F
)
}
\arguments{
\item{parfile}{An object of class MFCLPar}
}
\value{
An object of class numeric vector
}
\description{
Calculates recruitment periods for deterministic and stochastic projection settings
}
\examples{
data(par)
recPeriod(par)

}
