% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagTools.r
\name{recPeriod}
\alias{recPeriod}
\title{Recruitment Periods}
\usage{
recPeriod(
  par,
  af199 = NULL,
  af200 = NULL,
  pf232 = NULL,
  pf233 = NULL,
  show = FALSE
)
}
\arguments{
\item{par}{an object of class MFCLPar.}

\item{af199}{undocumented.}

\item{af200}{undocumented.}

\item{pf232}{undocumented.}

\item{pf233}{undocumented.}

\item{show}{undocumented.}
}
\value{
An object of class numeric vector.
}
\description{
Calculate recruitment periods for deterministic and stochastic projection
settings.
}
\examples{
data(par)
recPeriod(par)

}
