% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbMethods.R
\name{SBmethods}
\alias{SBmethods}
\alias{SB}
\alias{SB,MFCLRep,numeric,numeric-method}
\alias{SB,MFCLRep,missing,missing-method}
\alias{SB,MFCLRep,numeric,missing-method}
\alias{SB,MFCLRep,missing,numeric-method}
\alias{SB,list,numeric,numeric-method}
\alias{SB,list,missing,missing-method}
\alias{SB,list,numeric,missing-method}
\alias{SB,list,missing,numeric-method}
\alias{SBF0}
\alias{SBF0,MFCLRep,numeric,numeric-method}
\alias{SBF0,MFCLRep,missing,missing-method}
\alias{SBF0,MFCLRep,numeric,missing-method}
\alias{SBF0,MFCLRep,missing,numeric-method}
\alias{SBF0,list,numeric,numeric-method}
\alias{SBF0,list,missing,missing-method}
\alias{SBF0,list,numeric,missing-method}
\alias{SBF0,list,missing,numeric-method}
\alias{SBSBF0}
\alias{SBSBF0,MFCLRep,numeric,numeric,numeric,numeric-method}
\alias{SBSBF0,MFCLRep,missing,missing,missing,missing-method}
\alias{SBSBF0,list,numeric,numeric,numeric,numeric-method}
\alias{SBSBF0,list,missing,missing,missing,missing-method}
\alias{SBrecent}
\alias{SBrecent,MFCLRep-method}
\alias{SBrecent,list-method}
\alias{SBlatest}
\alias{SBlatest,MFCLRep-method}
\alias{SBlatest,list-method}
\alias{SBF0recent}
\alias{SBF0recent,MFCLRep-method}
\alias{SBF0recent,list-method}
\alias{SBSBF0recent}
\alias{SBSBF0recent,MFCLRep-method}
\alias{SBSBF0recent,list-method}
\alias{SBSBF0latest}
\alias{SBSBF0latest,MFCLRep-method}
\alias{SBSBF0latest,list-method}
\title{Methods for returns the spawning biomass (fished or unfished) and the depletion}
\usage{
SB(rep, mean_nyears, lag_nyears, ...)

\S4method{SB}{MFCLRep,numeric,numeric}(rep, mean_nyears, lag_nyears, combine_areas = TRUE)

\S4method{SB}{MFCLRep,missing,missing}(rep, mean_nyears, lag_nyears, ...)

\S4method{SB}{MFCLRep,numeric,missing}(rep, mean_nyears, lag_nyears, ...)

\S4method{SB}{MFCLRep,missing,numeric}(rep, mean_nyears, lag_nyears, ...)

\S4method{SB}{list,numeric,numeric}(rep, mean_nyears, lag_nyears, combine_areas = TRUE)

\S4method{SB}{list,missing,missing}(rep, mean_nyears, lag_nyears, ...)

\S4method{SB}{list,numeric,missing}(rep, mean_nyears, lag_nyears, ...)

\S4method{SB}{list,missing,numeric}(rep, mean_nyears, lag_nyears, ...)

SBF0(rep, mean_nyears, lag_nyears, ...)

\S4method{SBF0}{MFCLRep,numeric,numeric}(rep, mean_nyears, lag_nyears, combine_areas = TRUE)

\S4method{SBF0}{MFCLRep,missing,missing}(rep, mean_nyears, lag_nyears, ...)

\S4method{SBF0}{MFCLRep,numeric,missing}(rep, mean_nyears, lag_nyears, ...)

\S4method{SBF0}{MFCLRep,missing,numeric}(rep, mean_nyears, lag_nyears, ...)

\S4method{SBF0}{list,numeric,numeric}(rep, mean_nyears, lag_nyears, combine_areas = TRUE)

\S4method{SBF0}{list,missing,missing}(rep, mean_nyears, lag_nyears, ...)

\S4method{SBF0}{list,numeric,missing}(rep, mean_nyears, lag_nyears, ...)

\S4method{SBF0}{list,missing,numeric}(rep, mean_nyears, lag_nyears, ...)

SBSBF0(
  rep,
  sb_mean_nyears,
  sb_lag_nyears,
  sbf0_mean_nyears,
  sbf0_lag_nyears,
  ...
)

\S4method{SBSBF0}{MFCLRep,numeric,numeric,numeric,numeric}(
  rep,
  sb_mean_nyears,
  sb_lag_nyears,
  sbf0_mean_nyears,
  sbf0_lag_nyears,
  ...
)

\S4method{SBSBF0}{MFCLRep,missing,missing,missing,missing}(
  rep,
  sb_mean_nyears,
  sb_lag_nyears,
  sbf0_mean_nyears,
  sbf0_lag_nyears,
  ...
)

\S4method{SBSBF0}{list,numeric,numeric,numeric,numeric}(
  rep,
  sb_mean_nyears,
  sb_lag_nyears,
  sbf0_mean_nyears,
  sbf0_lag_nyears,
  ...
)

\S4method{SBSBF0}{list,missing,missing,missing,missing}(
  rep,
  sb_mean_nyears,
  sb_lag_nyears,
  sbf0_mean_nyears,
  sbf0_lag_nyears,
  ...
)

SBrecent(rep, ...)

\S4method{SBrecent}{MFCLRep}(rep, ...)

\S4method{SBrecent}{list}(rep, ...)

SBlatest(rep, ...)

\S4method{SBlatest}{MFCLRep}(rep, ...)

\S4method{SBlatest}{list}(rep, ...)

SBF0recent(rep, ...)

\S4method{SBF0recent}{MFCLRep}(rep, ...)

\S4method{SBF0recent}{list}(rep, ...)

SBSBF0recent(rep, ...)

\S4method{SBSBF0recent}{MFCLRep}(rep, ...)

\S4method{SBSBF0recent}{list}(rep, ...)

SBSBF0latest(rep, ...)

\S4method{SBSBF0latest}{MFCLRep}(rep, ...)

\S4method{SBSBF0latest}{list}(rep, ...)
}
\arguments{
\item{rep}{A single MFCLRep object or a list of MFCLRep objects.}

\item{mean_nyears}{The number of years to calculate the rolling mean over. A value of 1 is the same as not calculating a mean. For the SB() method the default value is 1.}

\item{lag_nyears}{The number of years to lag the output. A value of 0 is the same as not lagging the output. For the SB() method the default value is 0.}

\item{...}{Other arguments.}

\item{combine_areas}{If TRUE the biomassses in different areas are summed. Default value is TRUE.}

\item{sb_mean_nyears}{Equivalent to the argument \code{mean_nyears} for the SB part of the \code{SBSBF0()} method.}

\item{sb_lag_nyears}{Equivalent to the argument \code{lag_nyears} for the SB part of the \code{SBSBF0()} method.}

\item{sbf0_mean_nyears}{Equivalent to the argument \code{mean_nyears} for the SBF0 part of the \code{SBSBF0()} method.}

\item{sbf0_lag_nyears}{Equivalent to the argument \code{lag_nyears} for the SBF0 part of the \code{SBSBF0()} method.}
}
\value{
An FLQuant, or a list of FLQuant objects, depending on if the rep argument is a single MFCLRep object or a list.
}
\description{
Methods for returning the spawning biomass (\code{SB()}), unfished spawning biomass (\code{SBF0()}) and depletion (\code{SBSBF0()}) from a single MFCLRep object, or a list of MFCLRep objects.
The biomass is averaged over seasons.
It is possible to calculate a rolling mean over a specified number of years
as well as lag the output.
These can specified separately for the SB and SBF0 parts of SBSBF0.
The number of years in the outputs will match that of the inputs.
The start of the returned FLQuant(s) will be padded with NAs if necessary (for example, where the output has been lagged).
# When using lists of rep objects the lagging and rolling mean arguments are applied to all objects in the list.
}
\details{
Shortcut methods are available (\code{SBlatest()}, \code{SBrecent()}, \code{SBF0recent()} and \code{SBSBF0recent()}). These set the \code{mean_nyears()} and \code{lag_nyears()} parameters to match those used in the assessment reports.
\code{SBlatest()} does not calculate a mean or lag the output.
\code{SBrecent()} uses a rolling mean of 4 years without a lag.
\code{SBF0recent()} uses a rolling mean of 10 years with a lag of 1 year.
\code{SBSBF0recent()} is a combination of \code{SBrecent()} and \code{SBF0recent()}.
\code{SBSBF0latest()} is a combination of \code{SBlatest()} and \code{SBF0recent()}.
}
\examples{
\dontrun{
# Load up some example rep objects
rep1 <- read.MFCLRep(rep_file)
rep2 <- read.MFCLRep(rep_file)
rep3 <- read.MFCLRep(rep_file)
# SB 
SB(rep1) # equivalent to extracting the adultBiomass slot
SB(rep1, mean_nyears=4) # rolling mean with a window of 4 years
SBlatest(rep1) # no mean or lag
SBrecent(rep1) # Shortcut for a rolling mean of 4 years
# SBF0
SBF0(rep1) # equivalent to extracting the adultBiomass_nofish slot
SBF0(rep1, mean_nyears=10, lag_nyears=1) # a 10 year rolling window and a lag of 1
SBF0recent(rep1) # Shortcut for a 10 year rolling window and a lag of 1
# SBSBF0
SBSBF0(rep1) # equivalent to extracting dividing the adultBiomass slot by the adultBiomass_nofish slot
SBSBF0(rep1, sb_mean_nyears=4, sbf0_mean_nyears=10, sbf0_lag_nyears=1) # Depletion with SB with a 4 year rolling window and SBF0 with a 10 year rolling window and a lag of 1
SBSBF0recent(rep1) #  Shortcut for depletion with SB with a 4 year rolling window and SBF0 with a 10 year rolling window and a lag of 1
SBSBF0latest(rep1) #  Shortcut for depletion with SB without a rolling mean and no lag and SBF0 with a 10 year rolling window and a lag of 1
# Using lists of rep objects. The lagging and rolling mean arguments are applied to all objects in the list.
rep <- list(rep1=rep1, rep2=rep2, rep3=rep3)
SBSBF0recent(rep)
} 
}
